% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{acf.funts}
\alias{acf.funts}
\title{ACF for Functional Data}
\usage{
\method{acf}{funts}(x, ...)
}
\arguments{
\item{...}{Additional parameters to function based on data}
}
\value{
Functional ACF values
}
\description{
ACF for Functional Data
}
\examples{
acf(funts(electricity))
}
