% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{acf}
\alias{acf}
\alias{acf.default}
\alias{pacf}
\alias{pacf.default}
\title{Generic ACF/PACF}
\usage{
acf(x, ...)

\method{acf}{default}(x, ...)

pacf(x, ...)

\method{pacf}{default}(x, ...)
}
\arguments{
\item{...}{Additional parameters to function based on data}

\item{object}{Object for computation of (partial) autocorrelation function
(ACF/PACF)}
}
\value{
ACF or PACF values and plots
}
\description{
Generic ACF/PACF
}
\examples{
acf(1:10)
}
