% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArgMax.R
\name{Argmax}
\alias{Argmax}
\alias{int_x}
\title{Exact quantile calcuations for Confidence inervals}
\usage{
int_x(x, t1, t2)
}
\description{
Exact quantile calcuations for Confidence inervals
}
\keyword{internal}
