% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funts.R
\name{as.funts}
\alias{as.funts}
\title{Convert to funts}
\usage{
as.funts(x)
}
\arguments{
\item{x}{Object to convert to funts}
}
\value{
funts object
}
\description{
Convert to funts
}
\examples{
# as.funts(electricity)
}
