% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autodependence.R
\name{autocorrelation}
\alias{autocorrelation}
\title{Estimate the autocorrelation function of the series}
\usage{
autocorrelation(X, lags)
}
\arguments{
\item{X}{funts object of data of interest}

\item{lags}{Numeric(s) for the lags to estimate the lagged autocovariance / autocorrelation
operators.}
}
\value{
Return a list or data.frame with the lagged autocorrelation function(s)
estimated from the data. Each function is given by a \eqn{(r x r)}
matrix, where \eqn{r} is the number of points observed in each curve.
}
\description{
Obtain the empirical autocorrelation function for the given \code{lags} of a
functional time series, \code{X}. Given a functional time series, the sample
autocovariance functions \eqn{\hat{C}_{h}(u,v)} are given by:
\deqn{\hat{C}_{h}(u,v) =  \frac{1}{T} \sum_{i=1}^{T-h}(Y_{i}(u) -
     \overline{X}_{T}(u))(Y_{i+h}(v) - \overline{X}_{T}(v))}
where
\eqn{ \overline{X}_{T}(u) = \frac{1}{T} \sum_{i = 1}^{T} Y_{i}(t)}
denotes the sample mean function. The autocorrelation functions are defined
over the range \eqn{(0,1)} by normalizing these functions using the factor
\eqn{\int\hat{C}_{0}(u,u)du}.
}
\examples{
N <- 100
v <- seq(from = 0, to = 1, length.out = 10)
sig <- 2
bbridge <- generate_brownian_bridge(N, v, sig)

lagged_autocor <- autocorrelation(X = bbridge, lags = 0:1)
image(x = v, y = v, z = lagged_autocor$Lag0)
lagged_autocor1 <- autocorrelation(X = bbridge, lags = 1)
image(x = v, y = v, z = lagged_autocor1)
}
