% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facf_functions.R, R/tests_wn_ind_functions.R
\name{autocov_approx_h}
\alias{autocov_approx_h}
\title{Compute the approximate autocovariance at specified lag}
\usage{
autocov_approx_h(f_data, lag)

autocov_approx_h(f_data, lag)
}
\arguments{
\item{f_data}{the functional data matrix with observed functions in the columns}

\item{lag}{the lag to use to compute the single lag test statistic}
}
\value{
A 2-dimensional array encoding the autocovariance matrix for a given lag h.

A 2-dimensional array encoding the autocovariance matrix for a given lag h.
}
\description{
\code{autocov_approx_h} Computes the approximate autocovariance for a given lag h of the functional
data for every (t,s) in U_J X U_J.

\code{autocov_approx_h} Computes the approximate autocovariance for a given lag h of the functional
data
}
