% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfts.R
\name{average}
\alias{average}
\alias{mean.dfts}
\alias{median.dfts}
\title{Average Functions for dfts Objects}
\usage{
\method{mean}{dfts}(x, na.rm = TRUE, ...)

\method{median}{dfts}(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{A dfts object or data which can be automatically converted to that
format. See \code{\link[=dfts]{dfts()}}.}

\item{na.rm}{Boolean if NA values should be removed. Defaults to TRUE.}

\item{...}{Additional parameters to stats function}
}
\value{
Numeric vector
}
\description{
Compute the average values for dfts objects
}
\examples{
results <- mean(electricity)
results <- median(electricity)
}
