% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{bartlett_kernel}
\alias{bartlett_kernel}
\title{Bartlett Kernel}
\usage{
bartlett_kernel(l, h)
}
\arguments{
\item{l}{A numeric value indicating current lag}

\item{h}{A numeric value indicating smallest lag with no influence}
}
\value{
A numeric value indicating value of Kernel at that point
}
\description{
A very common Kernel. This is an obvious option in the Welsh-approximation.
}
\examples{
bartlett_kernel(1,2)
bartlett_kernel(0,0)
}
