% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_segmentation.R
\name{binary_segmentation}
\alias{binary_segmentation}
\title{Complete Binary Segmentation}
\usage{
binary_segmentation(
  X,
  statistic = c("Tn", "Mn"),
  method = c("Sim", "Approx", "Boot"),
  M = 20,
  J = 50,
  space = "BM",
  h = 1,
  iters = 1000,
  replace = FALSE,
  alpha = 0.05,
  final_verify = TRUE,
  silent = FALSE
)
}
\arguments{
\item{X}{Numeric data.frame with rows for evaluated values and columns
indicating functional observations}

\item{alpha}{Numeric value in \eqn{[0, 1]} indicating the significance for
cutoff_function.}

\item{final_verify}{(Optional) Boolean value XXXXXXXXXXXXX

Indicates if a final pass looking at sequences with only one change point
should be conducted to verify results. Note, this may modify existing
locations of change points, potentially to less accurate locations.}

\item{silent}{(Optional) Boolean value

Indicates if useful output should be silenced. Default FALSE shows output.}

\item{test_statistic_function}{XXXXXXXXXXXXXXXXX.}

\item{cutoff_function}{XXXXXX}

\item{...}{Additional arguments passed into test_statistic_function}
}
\value{
A list of numeric values indicating change points  (if exists),
NA otherwise
}
\description{
This function implements traditional binary segmentation on functional data
for general functions. Change points are recursively found until no
more change points are detected.
}
\examples{
binary_segmentation(X,"Tn","Boot")
}
