% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{cancer}
\alias{cancer}
\title{Breast Cancer}
\format{
\subsection{\code{cancer}}{

A data.frame with columns being the year and rows the age groups (5 years).
}
}
\usage{
cancer
}
\description{
Percentage of cause-specific deaths out of total deaths for female breast
cancer in the United States from 1950 to 2021.
}
\keyword{datasets}
