% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{change}
\alias{change}
\title{Change Function for funts Objects}
\usage{
change(
  X,
  method = c("characteristic", "mean", "robust", "eigenjoint", "eigensingle", "trace",
    "covariance", "pcamean", "pcadistribution"),
  statistic = c("Tn", "Mn"),
  critical = c("simulation", "permutation", "welch"),
  type = c("single", "segmentation", "elbow"),
  perm_type = "separate",
  replace = TRUE,
  max_changes = min(ncol(X), 20),
  CPs = NULL,
  blocksize = 1,
  d = 3,
  h = 3,
  M = 1000,
  J = 50,
  W = computeSpaceMeasuringVectors(X = X, M = 20, space = "BM"),
  K = bartlett_kernel,
  alpha = 0.05,
  cov.res = 30,
  weighting = 1/4,
  TVE = 0.95,
  trim_function = function(X) {
     0
 },
  errors = "L2",
  recommendation_change_points = 2,
  recommendation_improvement = 0.15,
  silent.binary = TRUE
)
}
\arguments{
\item{X}{Funts object or data easily convertible. See \code{\link[=funts]{funts()}}.}

\item{method}{Method to compute change point. Options include:
'characteristic', 'mean', 'robust', 'eigenjoint', 'eigensingle', 'trace', 'covariance',
'pcamean', and 'pcadistribution'.}

\item{statistic}{String for the Statistic. The integrated, \code{Tn}, or the
maximal, \code{Mn}.}

\item{critical}{String for method of computing threshold. Options are
'simulation', 'permutation', and 'welch'. However, Welch approximation is
not implemented for all methods.}

\item{type}{String for the type of change point detection, single change
('single'), binary segmentation ('segmentation'), or elbow plots ('elbow').}

\item{perm_type}{String indicating the type of permutation test to use.
Using 'separate' gives blocks which are separate while 'overlapping' creates
overlapping or sliding windows. When \code{blocksize=1} then these will be
identical.}

\item{replace}{Boolean for using a permutation or bootstrapped statistic when
\code{critical='permutation'}.}

\item{max_changes}{Integer as the max number of changes to search when using
type is \code{elbow}.}

\item{CPs}{Vector of change points to be given to the eigen test if the data
should be centered on these values first.}

\item{blocksize}{Integer for the width of the blocks when using a permutation
test. Can use \code{\link[=adaptive_bandwidth]{adaptive_bandwidth()}} is additional guidance is desired.}

\item{d}{Which eigenvalue or the number of eigenvalues which should be checked
in the eigenvalue tests.}

\item{h}{Number of lags used when computing long run covariance estimates. Used in
mean, characteristic, and eigenvalue tests.}

\item{M}{Number of simulations or permutations for critical values}

\item{J}{Resolution (J) in the characteristic method. The number of vectors
is defined by \code{W}.}

\item{W}{Space measuring vectors in characteristic method}

\item{K}{Kernel function for use in characteristic, mean, eigen, covariance
and pcamean}

\item{alpha}{Significance for Welch approximation}

\item{cov.res}{Resolution to use when computing covariance kernel changes}

\item{weighting}{Weights used in covariance kernel method and pcadistribution}

\item{TVE}{Total variance explained for pcamean}

\item{trim_function}{Trimming to be used in elbow plot}

\item{errors}{Type of errors used in elbow plot. Options are L2 and Trace}

\item{recommendation_change_points}{Number of lags forward to examine in deciding automated
elbow plot recommendation}

\item{recommendation_improvement}{Significant drop to look for in deciding automated elbow
plot recommendation}

\item{silent.binary}{Boolean if output should be printed when running binary
segmentation}
}
\value{
A list with
\enumerate{
\item pvalue:
\item location:
\item statistic:
\item simulations:
\item extra:
}
}
\description{
Change Function for funts Objects
}
\examples{
res <- change(electricity[,1:20],method='characteristic',critical = 'welch')
}
\references{
Aue, A., Rice, G., & Sonmez, O. (2018). Detecting and dating structural
breaks in functional data without dimension reduction. Journal of the Royal
Statistical Society. Series B, Statistical Methodology, 80(3), 509-529.
\url{https://doi.org/10.1111/rssb.12257}

Wegner, L., Wendler, M. Robust change-point detection for
functional time series based on U-statistics and dependent wild bootstrap.
Stat Papers (2024).

Aue, A., Rice, G., & Sonmez, O. (2020). Structural break
analysis for spectrum and trace of covariance operators. Environmetrics
(London, Ont.), 31(1)

Horvath, L., Rice, G., & Zhao, Y. (2022). Change point analysis
of covariance functions: A weighted cumulative sum approach. Journal of
Multivariate Analysis, 189, 104877-.

Berkes, I., Gabrys, R.,Horvath, L. & P. Kokoszka (2009).,
\emph{Detecting changes in the mean of functional observations}
Journal of the Royal Statistical Society, Series B 71, 927-946

Aue, A., Gabrys, R.,Horvath, L. & P. Kokoszka (2009).,
\emph{Estimation of a change-point in the mean function of functional data}
Journal of Multivariate Analysis 100, 2254-2269.

Huskova, M., & Meintanis, S.G. (2006). Change Point Analysis
based on Empirical Characteristic Functions. Metrika, 63, 145-168.
}
