% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_plot.R
\name{change_plot}
\alias{change_plot}
\title{Change Plot}
\usage{
change_plot(
  X,
  CPs,
  plot_title = NULL,
  val_axis_title = "Value",
  res_axis_title = "resolution",
  FD_axis_title = "FD Sims",
  eye = list(x = -1.5, y = -1.5, z = 1.5),
  aspectratio = list(x = 1, y = 1, z = 1),
  showticklabels = TRUE,
  warnings = TRUE
)
}
\arguments{
\item{CPs}{(Optional) Vectors of numeric values indicating the location of
change points. This will color each section differently. Default vallue
is NULL.}

\item{plot_title}{(Optional) String to title the plot. Default value is NULL.}

\item{val_axis_title}{(Optional) String to title the axis for values of
observation. Default value is 'Value'.}

\item{res_axis_title}{(Optional) String to title the axis for the resolution
range axis (observations in an FD object). Default value is 'resolution'.}

\item{FD_axis_title}{(Optional) String to title the axis for FD observations.
Default value is 'Observation'.}

\item{eye}{(Optional) List with certain parameters to determine the view of
the resulting image. The default value is list(x = -1.5, y = -1.5, z = 1.5).}

\item{aspectratio}{(Optional) List with certain parameters to determine the
image size / viewing parameters. The default value is
list(x=1, y=1, z=1), which is used in the non-interactive case. If
interactive is TRUE, then the default is c(2.5,.75,1) and all lists are
converted to this value so enter a vector if you wish to modify it for
the non-interactive case.}

\item{showticklabels}{(Optional) Boolean to indicate if the tick labels should
be given in the image. Default value is TRUE.}
}
\value{
An interactive plotly plot for the data.
}
\description{
This plots the functional data with means for each segment.
}
\examples{
change_plot(funts(electricity[,1:150]),CPs = c(66, 144))
#change_plot(funts(electricity),CPs = c(66, 144, 204, 243, 305))
}
