% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_detection.R
\name{characteristic_change_sim}
\alias{characteristic_change_sim}
\title{Characteristic Functions Change Point Detection - Simulation}
\usage{
characteristic_change_sim(
  X,
  M = 20,
  J = 50,
  nSims = 1000,
  h = 3,
  K = bartlett_kernel,
  space = "BM",
  silent = FALSE
)
}
\arguments{
\item{X}{Data.frame of the observations}

\item{M}{Numeric for the number of vectors to explore space}

\item{J}{Numeric for the reslolution to check}

\item{nSims}{Numeric for the number of simulations for distributions}

\item{h}{Numeric for bandwidth}

\item{K}{Function for bandwidth}

\item{space}{String for the space to explore space}

\item{silent}{Boolean to display outputs when running}
}
\value{
A list with the following elements:
'Tn': List with ('pval') p-value, ('gamProcess') gamma process using in
processes, and ('value') Tn statistic for data.
'Mn': List with ('pval') p-value, ('gamProcess') gamma process using in
processes, and ('value') Mn statistic for data.
}
\description{
Characteristic Functions Change Point Detection - Simulation
}
\examples{
characteristic_change_sim(electricity[,1:30],M = 10, J=25)
}
