% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_detection.R
\name{characteristic_change_sim_Tn}
\alias{characteristic_change_sim_Tn}
\title{Characteristic Functions Change Point Detection - Tn Simulation}
\usage{
characteristic_change_sim_Tn(
  X,
  M = 20,
  J = 50,
  nSims = 1000,
  h = 3,
  K = bartlett_kernel,
  space = "BM",
  silent = FALSE
)
}
\arguments{
\item{X}{Data.frame of the observations}

\item{M}{Numeric for the number of vectors to explore space}

\item{J}{Numeric for the reslolution to check}

\item{nSims}{Numeric for the number of simulations for distributions}

\item{h}{Numeric for bandwidth}

\item{K}{Function for bandwidth}

\item{space}{String for the space to explore space}

\item{silent}{Boolean to display outputs when running}
}
\value{
List with three items:
'pval': p-value of change point
'gamProcess': Simulated gamma processes
'value': Value of Tn test statistic for the data
}
\description{
Characteristic Functions Change Point Detection - Tn Simulation
}
\examples{
characteristic_change_sim_Tn(electricity[,1:50],M=5,J=10,h=0)
}
