% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_detection_welch.R
\name{characteristic_change_welch}
\alias{characteristic_change_welch}
\title{Characteristic Functions Change Point Detection - Welch Approximation}
\usage{
characteristic_change_welch(
  X,
  M = 20,
  J = 50,
  h = ncol(X$data)^(1/3),
  K = bartlett_kernel,
  space = "BM",
  alpha = 0.05
)
}
\arguments{
\item{X}{funts object}

\item{M}{Numeric for the number of vectors to measure space}

\item{J}{Numeric for the level of discretion}

\item{h}{Numeric for bandwidth}

\item{K}{Function for bandwidth}

\item{space}{String for the space to measure data}

\item{alpha}{Numeric for significance}
}
\value{
List with the following items:
'cutoff': Cutoff calculated for Tn using Welch approximation,
'value': Test statistic Tn for the data
'detected': Boolean if Tn is significant or not
'alpha': Significance of the result
}
\description{
Characteristic Functions Change Point Detection - Welch Approximation
}
\examples{
characteristic_change_welch(funts(electricity))
}
