% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_interval.R
\name{compute_confidence_interval}
\alias{compute_confidence_interval}
\title{Change Point Confidence Intervals}
\usage{
compute_confidence_interval(
  X,
  CPs,
  K = bartlett_kernel,
  h = 3,
  kappa = 0.5,
  M = 5000,
  alpha = 0.1,
  method = "distribution"
)
}
\arguments{
\item{X}{funts data}

\item{CPs}{Numeric vector for change points}

\item{K}{Default is bartlett_kernel.}

\item{h}{Default is 3.}

\item{kappa}{Default is 0.5.}

\item{M}{Default is 1000.}

\item{alpha}{Default is 0.05}

\item{method}{Options are 'distribution' and 'simulation'. Default is distribution.}
}
\value{
Data.frame with the first column for the change, second for the lower
confidence interval, and the third for the upper confidence interval.
}
\description{
Change Point Confidence Intervals
}
\examples{
X <- cbind(generate_brownian_motion(100,v=seq(0,1,0.05))$data,
           generate_brownian_motion(100,v=seq(0,1,0.05))$data+1000)
compute_confidence_interval(X,CPs = 100)
compute_confidence_interval(X,CPs=100,method = 'simulation')

X <- cbind(generate_brownian_motion(100,v=seq(0,1,0.05))$data,
           generate_brownian_motion(100,v=seq(0,1,0.05))$data+0.5)
compute_confidence_interval(X,100,alpha = 0.1)
compute_confidence_interval(X,CPs=100,alpha = 0.1,method = 'simulation')

X <- generate_brownian_motion(200,v=seq(0,1,0.05))
compute_confidence_interval(X,100)
compute_confidence_interval(X,100,method = 'simulation')

X <- cbind(generate_brownian_motion(200,v=seq(0,1,0.05))$data,
           generate_brownian_motion(100,v=seq(0,1,0.05))$data+0.1,
           generate_brownian_motion(150,v=seq(0,1,0.05))$data-0.05)
compute_confidence_interval(X,c(200,300))

# set.seed(12345)
# bs <- binary_segmentation(electricity,statistic = 'Tn',method = 'Boot')
compute_confidence_interval(X = electricity, CPs = c(66, 144, 204, 243, 305),alpha = 0.1)
}
\references{
Horváth, L., & Rice, G. (2024). Change Point Analysis for Time
Series (First edition.). Springer. \url{https://doi.org/10.1007/978-3-031-51609-2}

Aue, A., Rice, G., & Sönmez, O. (2018). Detecting and dating structural
breaks in functional data without dimension reduction. Journal of the Royal
Statistical Society. Series B, Statistical Methodology, 80(3), 509–529.
\url{https://doi.org/10.1111/rssb.12257}
}
