% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpss_test.R
\name{compute_kpss}
\alias{compute_kpss}
\title{Functional KPSS Test}
\usage{
compute_kpss(
  X,
  method = "MC",
  boot_method = "seperate",
  M = 1000,
  h = 3,
  TVE = 0.95,
  replace = TRUE
)
}
\arguments{
\item{X}{description}

\item{method}{description}

\item{boot_method}{description}

\item{M}{description}

\item{h}{description}

\item{TVE}{description}

\item{replace}{description}
}
\value{
List with statistic and p-value, respectively
}
\description{
Functional KPSS Test
}
\examples{
compute_kpss(generate_brownian_motion(100,v=seq(0,1,length.out=20)))
compute_kpss(generate_brownian_motion(100,v=seq(0,1,length.out=20)),
             method="BS")
compute_kpss(generate_brownian_motion(100,v=seq(0,1,length.out=20)),
             method="BS", boot_method='overlapping')
}
\references{
Chen, Y., & Pun, C. S. (2019). A bootstrap-based KPSS test for
functional time series. Journal of Multivariate Analysis, 174, 104535.

Kokoszka, P., & Young, G. (2016). KPSS test for functional time
series. Statistics, 50(5), 957-973.
}
