% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_change.R
\name{compute_mean_cutoff}
\alias{compute_mean_cutoff}
\title{Compute CUSUM Mean Statistic Cutoff}
\usage{
compute_mean_cutoff(data, alpha, h = 0, K = bartlett_kernel, M = 1000, ...)
}
\arguments{
\item{data}{funts object or numeric data.frame with evaled points on rows and fd objects in columns}

\item{alpha}{Numeric value indicating significance}

\item{h}{(Optional) The window parameter parameter for the estimation of the
long run covariance kernel. The default value is \code{h=0}, i.e., it
assumes iid data}

\item{K}{(Optional) Function indicating the Kernel to use if h>0}

\item{M}{(Optional) Number of Monte Carlo simulations used to get the critical
values. The default value is \code{M=1000}}

\item{...}{Unused, just for use in other functions}
}
\value{
Numeric indicating the cutoff value for CUSUM mean statistic
}
\description{
This function computes the cutoff for CUSUM statistic of a mean change
}
\examples{
\dontrun{
# Null Example
data_KL <- generate_data_fd(
  ns = c(100, 100),
  eigsList = list(
    c(3, 2, 1, 0.5),
    c(3, 2, 1, 0.5)
  ),
  basesList = list(
    fda::create.bspline.basis(nbasis = 4, norder = 4),
    fda::create.bspline.basis(nbasis = 4, norder = 4)
  ),
  meansList = c(0, 0),
  distsArray = c("Normal", "Normal"),
  evals = seq(0, 1, 0.05),
  kappasArray = c(0, 0)
)

compute_mean_cutoff(data_KL$data, 0.05)

# Mean CP Example
data_KL <- generate_data_fd(
  ns = c(100, 100),
  eigsList = list(
    c(3, 2, 1, 0.5),
    c(3, 2, 1, 0.5)
  ),
  basesList = list(
    fda::create.bspline.basis(nbasis = 4, norder = 4),
    fda::create.bspline.basis(nbasis = 4, norder = 4)
  ),
  meansList = c(0, 0.2),
  distsArray = c("Normal", "Normal"),
  evals = seq(0, 1, 0.05),
  kappasArray = c(0, 0)
)

compute_mean_cutoff(data_KL$data, 0.05)
}
}
