% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_change.R
\name{compute_mean_stat}
\alias{compute_mean_stat}
\title{#' Center Data
#'
#' This (internal) function centers each time point.
#'
#' This is an internal function, see use in .estimateCeps.
#'
#' @param data Numeric data.frame with evaled points on rows and fd objects in columns
#'
#' @return Numeric data.frame of the data, but with centered values at each time point
#'
#' @noRd
.centerData <- function(data) {
data - rowMeans(data)
}
Compute CUSUM Statistic for Mean Change}
\usage{
compute_mean_stat(data, k, weight = 0.5)
}
\arguments{
\item{data}{funts object or numeric data.frame with evaled points on rows and fd objects in columns}

\item{k}{Numeric indicating the change point of interest}

\item{...}{Unused, just for use in other functions}
}
\value{
Numeric indicating the CUSUM value at the given K value
}
\description{
This function is used to compute the CUSUM statistic for a mean change of
FD observations.
}
\examples{
\dontrun{
# Null Example
data_KL <- generate_data_fd(
  ns = c(100, 100),
  eigsList = list(
    c(3, 2, 1, 0.5),
    c(3, 2, 1, 0.5)
  ),
  basesList = list(
    fda::create.bspline.basis(nbasis = 4, norder = 4),
    fda::create.bspline.basis(nbasis = 4, norder = 4)
  ),
  meansList = c(0, 0),
  distsArray = c("Normal", "Normal"),
  evals = seq(0, 1, 0.05),
  kappasArray = c(0, 0)
)

compute_mean_stat(data_KL$data, 100)

# Mean CP Example
data_KL <- generate_data_fd(
  ns = c(100, 100),
  eigsList = list(
    c(3, 2, 1, 0.5),
    c(3, 2, 1, 0.5)
  ),
  basesList = list(
    fda::create.bspline.basis(nbasis = 4, norder = 4),
    fda::create.bspline.basis(nbasis = 4, norder = 4)
  ),
  meansList = c(0, 0.2),
  distsArray = c("Normal", "Normal"),
  evals = seq(0, 1, 0.05),
  kappasArray = c(0, 0)
)

compute_mean_stat(data_KL$data, 100)
}
}
