% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistics.R
\name{compute_Mn}
\alias{compute_Mn}
\title{Compute Mn Test Statistic}
\usage{
compute_Mn(
  X,
  k = NULL,
  M = 10000,
  W = NULL,
  space = "BM",
  which.Mn = FALSE,
  ...
)
}
\arguments{
\item{X}{Numeric data.frame with rows for evaluated values and columns
indicating FD.}

\item{k}{(Optional) Integer indicating if Defaults to NULL, but can be an
integer in \eqn{[1, ncol(X)]}.

Indicates the candidate change point to investigate (uses \code{compute_Mn()}
for selection). If k=NULL then get the test statistic for entire sample.
Default is NULL.}

\item{M}{(Optional) Numeric indicating the number of vectors used to span W.
Defaults to 10,000.}

\item{W}{(Optional) Data.frame of numerics with rows for evaluated values
and columns indicating function that contains the vectors to span the
space. Note the number of columns should match that of X. Default is NULL,
which generates the vectors in function. This overrides space and M if
not NULL.}

\item{space}{(Optional) String indicating the space to integrate against.
Default is "BM", or Brownian motion. See \code{computeSpaceMeasuringVectors()}
for more information.}

\item{which.Mn}{(Optional) Boolean which indicates if the location or test
statistic is of interest.}

\item{...}{Parameters are not passed anywhere. Just added for use in other
functions.}
}
\value{
A numeric value for the test statistic for entire sample or candidate
change point.
}
\description{
Function to calculate the change point test statistic for functional data
\deqn{Mn = sup_{x\in (0, 1)} \int |Z_n(v,x)|^2 dQ(v)}
}
\examples{
compute_Mn(electricity)
}
