% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationarity_test.R
\name{compute_stationarity_test}
\alias{compute_stationarity_test}
\title{Functional Stationarity Test}
\usage{
compute_stationarity_test(
  X,
  statistic = "Tn",
  method = "MC",
  boot_method = "seperate",
  M = 1000,
  h = 3,
  TVE = 0.95,
  replace = TRUE
)
}
\arguments{
\item{X}{The functional time series being tested, inputted in a matrix form
with each row representing each observation of the functional data values
on equidistant points of any prespecified interval.}

\item{statistic}{description}

\item{method}{description}

\item{boot_method}{description}

\item{M}{description}

\item{h}{description}

\item{TVE}{description}

\item{replace}{description}
}
\value{
The result of the test is presented with the value of test statistic and its p-value under the null hypothesis of local stationarity.
}
\description{
Stationarity test for functional time series with different methods on
determining the critical values of the test statistic. The Monte Carlo
method was constructed in Horvath et al. (2014), while the bootstrap-based
methods have not been validated in the literature (although such an option
is provided, please use them at your own risk).
}
\examples{
res <- compute_stationarity_test(
  generate_brownian_motion(100,v=seq(0,1,length.out=20)))
res <- compute_stationarity_test(
  generate_brownian_motion(100,v=seq(0,1,length.out=20)),statistic='Mn')
res1 <- compute_stationarity_test(
  generate_brownian_motion(1000,v=seq(0,1,length.out=30)))
res2 <- compute_stationarity_test(electricity)
}
\references{
Horvath, L., Kokoszka, P., & Rice, G. (2014). Testing
stationarity of functional time series. Journal of Econometrics,
179(1), 66-82.
}
