% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_detection.R
\name{compute_Tn}
\alias{compute_Tn}
\title{Compute Characteristic Integrated Test Statistic}
\usage{
compute_Tn(
  X,
  W = computeSpaceMeasuringVectors(M = 20, X = X, space = "BM"),
  J = 50
)
}
\arguments{
\item{X}{Data.frame of the observations}

\item{J}{Numeric for the reslolution to check}
}
\value{
Numeric integrated test statistic
}
\description{
Computes the integrated test statistic for the characteristic functional-based
change point detection.
}
\examples{
compute_Tn(electricity[,1:100])
}
