% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_measuring_vectors.R
\name{computeSpaceMeasuringVectors}
\alias{computeSpaceMeasuringVectors}
\title{Compute Spacing Measuring Vectors}
\usage{
computeSpaceMeasuringVectors(M, space, X)
}
\arguments{
\item{M}{Integer for the number of vectors to generate}

\item{space}{String for the space of interest. Options are Brownian motion
('BM'), OU process ('OU'), principal components ('PC'), and a vectors in iid
standard, random normals ('RN').}

\item{X}{Data of interest. Used to know evaluation amount of vectors}
}
\value{
Data.frame with columns of vectors describing the space. Columns are
independent vectors.
}
\description{
This function is used to compute vectors from spaces to measure the function
data objects.
}
\examples{
computeSpaceMeasuringVectors(10, "BM", electricity)
computeSpaceMeasuringVectors(10, "OU", electricity)
computeSpaceMeasuringVectors(10, "PC", electricity)
}
