% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_change.R
\name{cov_change}
\alias{cov_change}
\title{Covariance Change}
\usage{
cov_change(X, kappa = 1/4, len = 30)
}
\arguments{
\item{X}{Functional data}

\item{kappa}{XXXX}

\item{len}{XXXX}
}
\value{
XXXXX
}
\description{
This method implements a method for detection of covariance changes in
funtional data.
}
\examples{
# XXXXX
}
\references{
Aue, Alexander, et al. “Detecting and Dating Structural Breaks in Functional Data without Dimension Reduction.” Journal of the Royal Statistical Society. Series B, Statistical Methodology, vol. 80, no. 3, 2018, pp. 509–529, https://doi.org/10.1111/rssb.12257.
}
