% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_change.R
\name{criticalvalueMC}
\alias{criticalvalueMC}
\title{Compute critical values for TNstat (T_N)}
\usage{
criticalvalueMC(xf, len)
}
\arguments{
\item{xf}{XXXX}

\item{len}{XXXX}
}
\value{
XXXX
}
\description{
Compute critical values for TNstat (T_N)
}
\examples{
# XXXXX
}
