% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{cumsum.dfts}
\alias{cumsum.dfts}
\title{Cumsum for dfts object}
\usage{
\method{cumsum}{dfts}(x)
}
\arguments{
\item{x}{A dfts object or data which can be automatically converted to that
format. See \code{\link[=dfts]{dfts()}}.}
}
\value{
dfts object with data as cumsum
}
\description{
Cumsum for dfts object
}
\examples{
cumsum(electricity)
}
