% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{cumsum.funts}
\alias{cumsum.funts}
\title{Cumsum for funts object}
\usage{
\method{cumsum}{funts}(x)
}
\arguments{
\item{x}{funts object}
}
\value{
funts object with data as cumsum
}
\description{
Cumsum for funts object
}
\examples{
cumsum(funts(electricity))
}
