% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_changepoint.R
\name{detect_changepoint_singleCov}
\alias{detect_changepoint_singleCov}
\title{Estimate null and detect change point based on a single covar matrix}
\usage{
detect_changepoint_singleCov(
  X,
  nSims = 2000,
  x = seq(0, 1, length.out = 40),
  h = 3,
  K = bartlett_kernel,
  space = "BM",
  silent = FALSE,
  TN_M = 10000,
  Cov_M = 75
)
}
\arguments{
\item{X}{XXXXXX}

\item{nSims}{XXXXXX}

\item{x}{XXXXXX}

\item{h}{XXXXXX}

\item{K}{XXXXXX}

\item{space}{XXXXXX}

\item{silent}{XXXXXX}

\item{TN_M}{XXXXXX}

\item{Cov_M}{XXXXXX}
}
\value{
XXXXXX
}
\description{
Estimate null and detect change point based on a single covar matrix
}
\examples{
\dontrun{
X <- generate_data_fd(ns = c(100,100),
                      eigsList = list(c(3,2,1,0.5),c(3,2,1,0.5)),
                      basesList = list(fda::create.bspline.basis(nbasis=4, norder=4),
                                       fda::create.bspline.basis(nbasis=4, norder=4)),
                      meansList = c(0,1),
                      distsArray = c('Normal'),
                      evals = seq(0,1,0.05),
                      kappasArray = c(0.5),silent = TRUE)
cp_res <- detect_changepoint_singleCov(X, nSims=500, x=seq(0,1,length.out=20),
                                       h=3, K=bartlett_kernel, silent=FALSE)

X1 <- generate_data_fd(ns = c(200),
                       eigsList = list(c(3,2,1,0.5)),
                       basesList = list(fda::create.bspline.basis(nbasis=4, norder=4)),
                       meansList = c(0),
                       distsArray = c('Normal'),
                       evals = seq(0,1,0.05),
                       kappasArray = c(0),silent = TRUE)
nocp_res <- detect_changepoint_singleCov(X1, nSims=500, x=seq(0,1,length.out=20),
                                         h=0, K=bartlett_kernel, silent=FALSE)
X3 <- generate_data_fd(ns = c(50,250),
                       eigsList = list(c(3,2,1,0.5),c(30,1)),
                       basesList = list(fda::create.bspline.basis(nbasis=4, norder=4),
                                        fda::create.bspline.basis(nbasis=2, norder=2)),
                       meansList = c(0,0),
                       distsArray = c('Normal'), kappasArray = c(0.5),
                       evals = seq(0,1,0.05), silent = TRUE)
cp_res3 <- detect_changepoint_singleCov(X, nSims=500,
                                        x=seq(0,1,length.out=20))
}
}
