% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welch_approximation.R
\name{detect_changepoint_Welch}
\alias{detect_changepoint_Welch}
\title{Detect the Changepoint Based on Welch Approximation}
\usage{
detect_changepoint_Welch(
  X,
  M = 20,
  J = 50,
  h = ncol(X)^(1/3),
  K = bartlett_kernel,
  space = "BM",
  alpha = 0.05
)
}
\arguments{
\item{X}{Data.frame of the data}

\item{M}{Numeric for the number of vectors to measure space}

\item{J}{Numeric for the level of discretion}

\item{h}{Numeric for bandwidth}

\item{K}{Function for bandwidth}

\item{space}{String for the space to measure data}

\item{alpha}{Numeric for significance}
}
\description{
Detect the Changepoint Based on Welch Approximation
}
