% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfts.R
\name{dfts_group}
\alias{dfts_group}
\alias{Math.dfts}
\alias{Ops.dfts}
\title{Group Generic Functions}
\usage{
\method{Math}{dfts}(x, ...)

\method{Ops}{dfts}(e1, e2)
}
\arguments{
\item{x, e1, e2}{dfts object}

\item{...}{Further arguments passed to the methods}
}
\value{
dfts object
}
\description{
Group generic methods defined for things like Math, Ops, and so forth.
}
\examples{
result <- sqrt( electricity )
result <- electricity + electricity
result1 <- electricity * electricity
}
