% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facf_functions.R, R/tests_wn_ind_functions.R
\name{diagonal_covariance_i}
\alias{diagonal_covariance_i}
\title{Diagonal Covariance}
\usage{
diagonal_covariance_i(f_data, i)

diagonal_covariance_i(f_data, i)
}
\arguments{
\item{f_data}{the functional data matrix with observed functions in the columns}

\item{i}{the index in 1:T that we are computing the covariance for}
}
\value{
A 2-dimensional array, encoding the covariance matrix of the functional data for lag
`windows defined by i.

A 2-dimensional array, encoding the covariance matrix of the functional data for lag
windows defined by i.
}
\description{
Computes the approximate diagonal covariance matrix of the functional
data for lag windows defined by i.

c_f_data <- center(f_data)
sum_parts <- as.list((1+max(i,j)):N)
sum_parts <- lapply(sum_parts,
function(k) c_f_data\link{,k-i} \%o\% c_f_data\link{,k} \%o\%
c_f_data\link{,k-j} \%o\% c_f_data\link{,k})
}
\details{

}
