% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfts.R
\name{dim.dfts}
\alias{dim.dfts}
\title{Dimension of dfts Object}
\usage{
\method{dim}{dfts}(x, ...)
}
\arguments{
\item{x}{A dfts object or data which can be automatically converted to that
format. See \code{\link[=dfts]{dfts()}}.}

\item{...}{Additional parameters to stats function}
}
\value{
Numerics indicating the dimension of the dfts object
}
\description{
Retrieve the dimension of a dfts object.
}
\examples{
results <- dim(electricity)
}
