% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funts.R
\name{dim.funts}
\alias{dim.funts}
\title{Dimension of funts Object}
\usage{
\method{dim}{funts}(x, ...)
}
\arguments{
\item{x}{funts object}

\item{...}{Additional parameters to stats function}
}
\value{
Numerics indicating the dimension of the funts object
}
\description{
Retrieve the dimension of a funts object.
}
\examples{
results <- dim(funts(electricity))
}
