% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq.R
\name{distplot}
\alias{distplot}
\title{Distribution plot}
\usage{
distplot(
  X,
  CPs = NULL,
  TVE = 0.95,
  max.d = 3,
  distribution = "norm",
  qq.sep = TRUE,
  alpha = 0.05,
  legend = FALSE,
  ...
)
}
\arguments{
\item{CPs}{description}

\item{TVE}{description}

\item{max.d}{description}

\item{distribution}{description}

\item{qq.sep}{description}

\item{alpha}{description}

\item{legend}{description}

\item{...}{description}

\item{x}{description}
}
\value{
ggplot2 for plot to compare principal components to a distribution
}
\description{
Distribution plot
}
\examples{
result <- distplot(electricity)
result1 <- distplot(electricity, max.d=2, qq.sep=FALSE)
result2 <- distplot(electricity,dist='exp')
}
\references{
John Fox, & Sanford Weisberg (2019). An R Companion to Applied
Regression. Sage.
}
