% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{distribution_plot}
\alias{distribution_plot}
\title{Distribution Plot}
\usage{
distribution_plot(object, CPs = NULL, alpha = 0.05)
}
\arguments{
\item{object}{funts object}

\item{CPs}{CPs, if there are any. Default is NULL.}
}
\value{
Plot (ggplot2) with colored observations / regions
}
\description{
Plot functional data as 2-dimensional data with the mean(s) and the related
distribution(s) given.
}
\examples{
distribution_plot(funts(electricity))
distribution_plot(funts(electricity), CPs=c(50,100,200))
}
