% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoElbow_Method.R
\name{.autoElbow_method}
\alias{.autoElbow_method}
\title{Auto Elbow Method}
\usage{
.autoElbow_method(
  data,
  test_stat_method = compute_Mn,
  trim_function_method = trim_function,
  newVarExplain = 0.05,
  maxCP = 10,
  errorType = "L2",
  demean = TRUE,
  ...
)
}
\arguments{
\item{data}{Numeric data.frame with rows for evaluated values and columns
indicating FD}

\item{test_stat_method}{XXXXX}

\item{trim_function_method}{Function taking data as an argument and returning a
numeric value indicating how much should be trimmed on each end}

\item{newVarExplain}{XXXXXXXXX}

\item{maxCP}{XXXXXXXXX}

\item{errorType}{String of 'L2' or 'Tr' indicating the error function to use}

\item{demean}{XXXXXXXXX}

\item{...}{Additional parameters to pass into the respective functions}
}
\value{
list with element 1 the data frame with the change point location and
element a ggplot of variance as a function of CPs
}
\description{
Method to determine the number of change points using the elbow method. Note,
cascading change points are not considered to allow for every possible
number of change points to be selectable.
}
\examples{
\dontrun{
data_KL <- generate_data_fd(ns = c(100,50,100),
                  eigsList = list(c(3,2,1,0.5),
                                  c(3,2,1,0.5),
                                  c(3,2,1,0.5)),
                  basesList = list(
                     fda::create.bspline.basis(nbasis=4, norder=4),
                     fda::create.bspline.basis(nbasis=4, norder=4),
                     fda::create.bspline.basis(nbasis=4, norder=4)),
                  meansList = c(-1,1,-1),
                  distsArray = c('Normal','Normal','Normal'),
                  evals = seq(0,1,0.05),
                  kappasArray = c(0, 0, 0))

results <- .autoElbow_method(data=data_KL,
                              test_stat_method=compute_Tn)
}
}
