% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_segmentation.R
\name{.ce_bootstrap}
\alias{.ce_bootstrap}
\title{Characteristic Detection - Mn Method}
\usage{
.ce_bootstrap(
  X,
  statistic = c("Tn", "Mn"),
  W = computeSpaceMeasuringVectors(M, space, X),
  J = 50,
  space = "BM",
  blockSize = 1,
  iters = 1000,
  replace = FALSE,
  alpha = 0.05,
  silent = FALSE
)
}
\arguments{
\item{X}{numeric matrix of fd}

\item{statistic}{Test statistic of interest}

\item{W}{Space Measuring Vectors}

\item{silent}{(Optional) Boolean value

Indicates if useful output should be silenced. Default FALSE shows output.}
}
\value{
List with (1) value, (2) cutoff, (3) p-value, and (4) bootstrapped samples
}
\description{
Characteristic Detection - Mn Method
}
