% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_segmentation.R
\name{.ce_detect_Tn}
\alias{.ce_detect_Tn}
\title{Characteristic Detection - Tn Method}
\usage{
.ce_detect_Tn(
  X,
  W,
  M = 20,
  J = 50,
  nSims = 1000,
  h = 3,
  K = bartlett_kernel,
  silent = FALSE
)
}
\arguments{
\item{X}{numeric matrix of fd}

\item{W}{Space Measuring Vectors}

\item{nSims}{iters from before}

\item{K}{Kernel function}

\item{silent}{(Optional) Boolean value

Indicates if useful output should be silenced. Default FALSE shows output.}
}
\value{
List with p-value for change detection
}
\description{
Characteristic Detection - Tn Method
}
