% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_segmentation.R
\name{.ce_recursive_segmentation}
\alias{.ce_recursive_segmentation}
\title{Characteristic Change - Recursive Segmentation}
\usage{
.ce_recursive_segmentation(
  X,
  fn,
  M = 20,
  J = 50,
  space = "BM",
  h_function = function(X) {
     ncol(X)^(1/3)
 },
  iters = 1000,
  replace = FALSE,
  alpha = 0.05,
  silent = FALSE,
  addAmt = 0
)
}
\arguments{
\item{X}{numeric matrix of fd}

\item{fn}{Method for test statistic}

\item{alpha}{Numeric value in \eqn{[0, 1]} indicating the significance for
cutoff_function.}

\item{silent}{(Optional) Boolean value

Indicates if useful output should be silenced. Default FALSE shows output.}

\item{addAmt}{Number to add when recursively splitting to get correct
change location}
}
\value{
Numeric vector of changes
}
\description{
Characteristic Change - Recursive Segmentation
}
