% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_segmentation.R
\name{.ce_verify_changes}
\alias{.ce_verify_changes}
\title{Characteristic Change - Verify Segmentation}
\usage{
.ce_verify_changes(
  CPsVals,
  X,
  fn,
  M = M,
  J = J,
  space = space,
  h_function = function(X) {
     ncol(X)^(1/3)
 },
  iters = iters,
  replace = replace,
  alpha = alpha,
  silent = FALSE
)
}
\arguments{
\item{CPsVals}{Numeric vector of potential change points}

\item{X}{funts object or numeric data.frame with rows for evaluated values and columns
indicating functional observations}

\item{fn}{Method for test statistic}

\item{alpha}{Numeric value in \eqn{[0, 1]} indicating the significance for
cutoff_function.}

\item{silent}{(Optional) Boolean value

Indicates if useful output should be silenced. Default FALSE shows output.}
}
\description{
Characteristic Change - Verify Segmentation
}
