% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_data}
\alias{.check_data}
\title{Check data for functions}
\usage{
.check_data(X)
}
\arguments{
\item{X}{funts or object that can be directly converted into funts object}
}
\value{
funts object or error if the data is incorrect
}
\description{
Check data for functions
}
\examples{
#.check_data(funts(electricity))
#.check_data(electricity)
}
