% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facf.R
\name{.compute_autocovariance}
\alias{.compute_autocovariance}
\title{Estimate Autocovariance Function of funts}
\usage{
.compute_autocovariance(X, lag.max)
}
\arguments{
\item{X}{funts object}

\item{lag.max}{Number of lagged covariance operators
of the functional time series that will be used
to estimate the autocorrelation function. Default of NULL results in 20 lags.}
}
\value{
Return a list with the lagged autocovariance
functions estimated from the data. Each function is given
by a \eqn{(m x m)} matrix, where \eqn{m} is the
resolution observed in each curve.
}
\description{
Obtain the empirical autocovariance function for lags
\eqn{= 0,...,}\code{lag.max} of the functional time series. Given
\eqn{X_{1},...,X_{T}} a functional time series, the sample
autocovariance functions \eqn{\hat{C}_{h}(u,v)} are given by:
\deqn{\hat{C}_{h}(u,v) =  \frac{1}{T} \sum_{i=1}^{T-h}(Y_{i}(u) - \overline{Y}_{T}(u))(Y_{i+h}(v) - \overline{Y}_{T}(v))}
where
\eqn{ \overline{Y}_{T}(u) = \frac{1}{T} \sum_{i = 1}^{T} Y_{i}(t)}
denotes the sample mean function.
}
\examples{
.compute_autocovariance(funts(electricity), 10)
}
