% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facf.R
\name{.compute_FACF}
\alias{.compute_FACF}
\title{Obtain the autocorrelation function for a given functional time series.}
\usage{
.compute_FACF(
  X,
  lag.max = NULL,
  alpha = 0.05,
  method = c("Welch", "MC", "Imhof"),
  WWN = TRUE,
  figure = TRUE,
  ...
)
}
\arguments{
\item{X}{funts object}

\item{lag.max}{Number of lagged covariance operators
of the functional time series that will be used
to estimate the autocorrelation function. Default of NULL results in 20 lags.}

\item{alpha}{A value between 0 and 1 that indicates significant level for
the confidence interval for the i.i.d. bounds of the autocorrelation
function. By default \code{alpha = 0.95}.}

\item{method}{Character specifying the method to be used when estimating the
distribution under the hypothesis of functional white noise.
Accepted values are:
\itemize{
\item "Welch": Welch approximation.
\item "MC": Monte-Carlo estimation.
\item "Imhof": Estimation using Imhof's method.
}
By default, \code{method = "Welch"}.}

\item{WWN}{Logical. If \code{TRUE}, WWN bounds are also computed}

\item{figure}{Logical. If \code{TRUE}, plots the estimated autocorrelation
function with the specified bounds.}

\item{...}{Further arguments passed to the \code{.plot_FACF} function.}
}
\value{
Return a list with:
\itemize{
\item \code{SWN_bound}: The upper prediction
bound for the i.i.d. distribution under strong white noise assumption.
\item \code{WWN_bound}: The upper prediction
bound for the i.i.d. distribution under weak white noise assumption.
\item \code{acfs}: Autocorrelation values for
each lag of the functional time series.
}
}
\description{
Estimate the lagged autocorrelation function for a given
functional time series and its distribution under the
hypothesis of strong functional white noise. This graphic tool
can be used to identify seasonal patterns in the functional
data as well as auto-regressive or moving average terms.
Strong white noise is included and weak white noise can also be included to
test the presence of serial correlation in the data.
}
\examples{
# Example 1

N <- 100
v <- seq(from = 0, to = 1, length.out = 5)
sd <- 2
X <- generate_brownian_bridge(N, v, sd)
.compute_FACF(X,20)

\donttest{
# Example 2

.compute_FACF(X=funts(electricity),lag.max = 30,alpha = 0.90)
}
}
\references{
Mestre G., Portela J., Rice G., Muñoz San Roque A., Alonso E. (2021).
\emph{Functional time series model identification and diagnosis by
means of auto- and partial autocorrelation analysis.}
Computational Statistics & Data Analysis, 155, 107108.
\url{https://doi.org/10.1016/j.csda.2020.107108}

Mestre, G., Portela, J., Muñoz-San Roque, A., Alonso, E. (2020).
\emph{Forecasting hourly supply curves in the Italian Day-Ahead
electricity market with a double-seasonal SARMAHX model.}
International Journal of Electrical Power & Energy Systems,
121, 106083. \url{https://doi.org/10.1016/j.ijepes.2020.106083}

Kokoszka, P., Rice, G., Shang, H.L. (2017).
\emph{Inference for the autocovariance of a functional
time series under conditional heteroscedasticity}
Journal of Multivariate Analysis,
162, 32--50. \url{https://doi.org/10.1016/j.jmva.2017.08.004}
}
