% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationarity_test.R
\name{.compute_stationary_test_stat}
\alias{.compute_stationary_test_stat}
\title{Title}
\usage{
.compute_stationary_test_stat(X, statistic)
}
\arguments{
\item{statistic}{}
}
\description{
Title
}
\keyword{internal}
