% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_detection_welch.R
\name{.compute_Welch}
\alias{.compute_Welch}
\title{Welch Approximation to Tn}
\usage{
.compute_Welch(
  X,
  alpha = 0.05,
  W = NULL,
  M = 100,
  J = 50,
  h = ncol(X$data)^(1/3),
  K = bartlett_kernel
)
}
\arguments{
\item{X}{Numeric data.frame with evaled points on rows and fd objects in columns}

\item{alpha}{(Optional) Numeric value in (0, 1) for significance. Default is 0.05}

\item{W}{(Optional) Data.frame for the functions to integrate against
used for both muHat and sigmaHat2. Default of NULL will use Brownian
motion.}

\item{M}{(Optional) Numeric value for number of functions to integrate against
(ncol(W)). Default is 100.}

\item{h}{(Optional) Numeric value indicating the amount of blocking. Default
is Tval^(1/3)}

\item{K}{(Optional) Function used for CHat, indicating Kernel. Default is the
Bartlett Kernel.}

\item{TVal}{(Optional) Numeric value indicate the number of FDs. Default is the
number of columns in X.}

\item{W1}{(Optional)  data.frame for the functions to integrate against
used for sigmaHat2. Default of NULL will use Brownian motion.}

\item{...}{Unused parameters, but added for use in other functions.}
}
\value{
Numeric value indicating cutoff from Welsh approximation
}
\description{
This function approximates the Tn statistic using the Welch approximation.
}
\examples{
.compute_Welch(electricity[,1:40], h = 0)
}
