% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facf.R
\name{.estimate_iid_distr_Imhof}
\alias{.estimate_iid_distr_Imhof}
\title{Estimate distribution of the fACF under the iid. hypothesis using Imhof's method}
\usage{
.estimate_iid_distr_Imhof(X, autocovs, l2norms)
}
\arguments{
\item{autocovs}{An \eqn{(m x m)} matrix with the discretized
values of the autocovariance operator \eqn{\hat{C}_{0}}, obtained by
calling the function \code{obtain_autocovariance}. The value
\eqn{m} indicates the number of points observed in each curve.}

\item{l2norms}{A vector containing the L2 norm of
the autocovariance function. It can be obtained by calling
function \code{obtain_suface_L2_norm}.}
}
\value{
Return a list with:
\itemize{
\item \code{ex}: Knots where the distribution has been estimated.
\item \code{ef}: Discretized values of the estimated distribution.
}
}
\description{
Estimate the distribution of the autocorrelation function
under the hypothesis of strong functional white noise. This
function uses Imhof's method to estimate the distribution.
}
\examples{
# Example 1

N <- 100
v <- seq(from = 0, to = 1, length.out = 10)
sig <- 2
Y <- generate_brownian_bridge(N, v, sig)
lag.max <- 1
autocovSurface <- obtain_autocovariance(Y,lag.max)
matindex <- .obtain_suface_L2_norm (v,autocovSurface)
# Remove lag 0
matindex <- matindex[-1]
Imhof_dist <- .estimate_iid_distr_Imhof(Y,autocovSurface,matindex)
plot(Imhof_dist$ex,Imhof_dist$ef,type = "l",main = "ecdf obtained by Imhof's method")
grid()

\donttest{
# Example 2

N <- 400
v <- seq(from = 0, to = 1, length.out = 50)
sig <- 2
Y <- generate_brownian_bridge(N, v, sig)
autocovSurface <- obtain_autocovariance(Y,lag.max)
matindex <- .obtain_suface_L2_norm (v,autocovSurface)
# Remove lag 0
matindex <- matindex[-1]
Imhof_dist <- .estimate_iid_distr_Imhof(Y,autocovSurface,matindex)
plot(Imhof_dist$ex,Imhof_dist$ef,type = "l",main = "ecdf obtained by Imhof's method")
grid()
}
}
