% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facf.R
\name{.estimate_iid_distr_MC}
\alias{.estimate_iid_distr_MC}
\title{Estimate distribution of the fACF under the iid hypothesis using MC method}
\usage{
.estimate_iid_distr_MC(X, autocovSurface, matindex, nsims = 10000)
}
\arguments{
\item{X}{funts object}

\item{autocovSurface}{An \eqn{(m x m)} matrix with the discretized
values of the autocovariance operator \eqn{\hat{C}_{0}}, obtained
by calling the function \code{obtain_autocovariance}.
The value \eqn{m} indicates the number of points observed
in each curve.}

\item{matindex}{A vector containing the L2 norm of
the autocovariance function. It can be obtained by calling
function \code{obtain_suface_L2_norm}.}

\item{nsims}{Positive integer indicating the number of
MC simulations that will be used to estimate the distribution
of the statistic. Increasing the number of simulations will
improve the estimation, but it will increase the computational
time.
By default, \code{nsims = 10000}.}
}
\value{
Return a list with:
\itemize{
\item \code{ex}: Knots where the distribution has been estimated.
\item \code{ef}: Discretized values of the estimated distribution.
\item \code{Reig}: Raw values for iid statistic of each MC simulation.
}
}
\description{
Estimate the distribution of the autocorrelation function under the
hypothesis of strong functional white noise. This function uses a
Monte Carlo method to estimate the distribution.
}
\examples{
# Example 1

N <- 100
v <- seq(from = 0, to = 1, length.out = 10)
sig <- 2
Y <- generate_brownian_bridge(N, v, sig)
lag.max <- 1
autocovSurface <- obtain_autocovariance(Y,lag.max)
matindex <- .obtain_suface_L2_norm(v,autocovSurface)
# Remove lag 0
matindex <- matindex[-1]
MC_dist <- .estimate_iid_distr_MC(Y,autocovSurface,matindex)
plot(MC_dist$ex,MC_dist$ef,type = "l",main = "ecdf obtained by MC simulation")
grid()

\donttest{
# Example 2

N <- 400
v <- seq(from = 0, to = 1, length.out = 50)
sig <- 2
Y <- generate_brownian_bridge(N, v, sig)
lag.max <- 20
autocovSurface <- obtain_autocovariance(Y,lag.max)
matindex <- .obtain_suface_L2_norm(v,autocovSurface)
# Remove lag 0
matindex <- matindex[-1]
MC_dist <- .estimate_iid_distr_MC(Y,autocovSurface,matindex)
plot(MC_dist$ex,MC_dist$ef,type = "l",main = "ecdf obtained by MC simulation")
grid()
}
}
