% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facf.R
\name{.fit_ARHp_FPCA}
\alias{.fit_ARHp_FPCA}
\title{Fit an ARH(p) to a given functional time series}
\usage{
.fit_ARHp_FPCA(X, p, n_pcs, show_varprop = TRUE)
}
\arguments{
\item{p}{Numeric value specifying the order
of the functional autoregressive
model to be fitted.}

\item{n_pcs}{Number of principal components
that will be used to fit the ARH(p) models.}

\item{show_varprop}{Logical. If \code{show_varprop = TRUE}, a plot of
the proportion of variance explained by the first \code{n_pcs} functional
principal components will be shown. By default \code{show_varprop = TRUE}.}
}
\description{
Fit an \eqn{ARH(p)} model to a given functional time series. The fitted
model is based on the model proposed in (Aue et al, 2015), first
decomposing the original functional observations into a vector time series
of \code{n_pcs} FPCA scores, and then fitting a vector autoregressive
model of order \eqn{p} (\eqn{VAR(p)}) to the time series of the scores.
Once fitted, the Karhunen-Loève expansion is used to re-transform the
fitted values into functional observations.
}
\examples{
# Example 1

N <- 250
dv <- 20
v <- seq(from = 0, to = 1, length.out = 20)

phi <- 1.3 * ((v) \%*\% t(v))

persp(v,v,phi,
      ticktype = "detailed",
      main = "Integral operator")

set.seed(3)
white_noise <-  generate_brownian_bridge(N, v = v)

y <- matrix(nrow = dv, ncol = N)
y[,1] <- white_noise$data[,1]
for(jj in 2:N){
    y[,jj] <- white_noise$data[,jj];

    y[,jj] <- y[,jj] + .integral_operator(operator_kernel = phi,
                                    v = v, curve = y[,jj-1])
}

# Fit an ARH(1) model
mod <- .fit_ARHp_FPCA(X = funts(y,intraobs = v), p = 1, n_pcs = 5)

# Plot results
plot(v, y[,50], type = "l", lty = 1, ylab = "")
lines(v, mod$x_est[,50], col = "red")
legend("bottomleft", legend = c("real","est"),
       lty = 1, col = c(1,2))

}
\references{
Aue, A., Norinho, D. D., Hormann, S. (2015).
\emph{On the Prediction of Stationary Functional Time Series}
Journal of the American Statistical Association,
110, 378--392. \url{https://doi.org/10.1080/01621459.2014.909317}
}
