% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_change.R
\name{.l2norm}
\alias{.l2norm}
\title{L2 Norm}
\usage{
.l2norm(vec)
}
\arguments{
\item{vec}{Vector of numerics to compute L2 norm}
}
\value{
Numeric L2 norm value
}
\description{
This (internal) function computes the L2 norm of the data.
}
\details{
See use in tau_est
}
