% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{.linear_imputatation}
\alias{.linear_imputatation}
\title{Linear Imputation}
\usage{
.linear_imputatation(data, obs_share_data = FALSE)
}
\arguments{
\item{data}{funts object}

\item{obs_share_data}{Boolean to indicate if the end of an observation
should inform the next. Default is FALSE.}
}
\value{
funts object with imputed values
}
\description{
Linear Imputation
}
