% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facf.R
\name{.obtain_autocov_eigenvalues}
\alias{.obtain_autocov_eigenvalues}
\title{Estimate eigenvalues of the autocovariance function}
\usage{
.obtain_autocov_eigenvalues(X, epsilon = 1e-04)
}
\arguments{
\item{X}{funts object}

\item{epsilon}{Value used to determine how many eigenvalues will be returned.
The eigenvalues \eqn{\lambda_{j}/\lambda_{1} > \code{epsilon}} will be
returned. By default \code{epsilon = 0.0001}.}
}
\value{
A vector containing the \eqn{k} eigenvalues
greater than \code{epsilon}.
}
\description{
Estimate the eigenvalues of the sample autocovariance
function \eqn{\hat{C}_{0}}. This functions returns the eigenvalues which
are greater than the value \code{epsilon}.
}
\examples{
N <- 100
v <- seq(from = 0, to = 1, length.out = 10)
sig <- 2
Y <- generate_brownian_bridge(N, v, sig)
lambda <- .obtain_autocov_eigenvalues(X = Y)
}
