% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facf.R
\name{.obtain_suface_L2_norm}
\alias{.obtain_suface_L2_norm}
\title{Obtain L2 norm of the autocovariance functions}
\usage{
.obtain_suface_L2_norm(intraobs, autocovs)
}
\arguments{
\item{intraobs}{Discretization points of the curves.}

\item{autocovs}{An \eqn{(m x m)} matrix with the discretized
values of the autocovariance operator \eqn{\hat{C}_{0}}, obtained by
calling the function \code{obtain_autocovariance}. The value
\eqn{m} indicates the number of points observed in each curve.}
}
\value{
A vector containing the L2 norm of the
lagged autocovariance functions \code{autocovs}.
}
\description{
Returns the L2 norm of the lagged autocovariance functions
\eqn{\hat{C}_{h}}. The L2 norm of these functions is defined as
\deqn{\sqrt(\int \int \hat{C}^{2}_{h}(u,v)du dv)}.
}
\examples{
# Example 1

N <- 100
v <- seq(from = 0, to = 1, length.out = 10)
sd <- 2
X <- generate_brownian_bridge(N, v, sd)
lag.max <- 1
autocovSurface <- obtain_autocovariance(X=X,nlags = lag.max)
norms <- .obtain_suface_L2_norm(intraobs = v,autocovs = autocovSurface)
plot_autocovariance(fun.autocovariance = autocovSurface,lag = 1)
title(sub = paste0("Lag ",1," - L2 Norm: ",norms[2]))

\donttest{
# Example 2

N <- 400
v <- seq(from = 0, to = 1, length.out = 50)
sig <- 2
X <- generate_brownian_bridge(N, v, sig)
lag.max <- 2
autocovSurface <- obtain_autocovariance(X=X,nlags = lag.max)
norms <- .obtain_suface_L2_norm(intraobs = v,autocovs = autocovSurface)
opar <- par(no.readonly = TRUE)
par(mfrow = c(1,3))
plot_autocovariance(fun.autocovariance = autocovSurface,lag = 0)
title(sub = paste0("Lag ",0," - L2 Norm: ",norms[1]))
plot_autocovariance(fun.autocovariance = autocovSurface,lag = 1)
title(sub = paste0("Lag ",1," - L2 Norm: ",norms[2]))
plot_autocovariance(fun.autocovariance = autocovSurface,lag = 2)
title(sub = paste0("Lag ",2," - L2 Norm: ",norms[3]))
par(opar)
}
}
