% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_examination.R
\name{.pca_mult}
\alias{.pca_mult}
\title{Principal Component Multiplication}
\usage{
.pca_mult(pca, pc_num = 1)
}
\arguments{
\item{pca}{PCA object from pca}

\item{pc_num}{PC of interest}
}
\value{
pca coefficient from the pca and its rotation
}
\description{
Principal Component Multiplication
}
\examples{
vals <- pca(funts(electricity))
res <- .pca_mult(vals,1)
}
