% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_examination.R
\name{.pca_reconstruct}
\alias{.pca_reconstruct}
\title{PCA reconstruction}
\usage{
.pca_reconstruct(pca, new_rot = NULL)
}
\arguments{
\item{pca}{PCA object from pca}

\item{new_rot}{Forecasted rotation}
}
\value{
Data.frame of forecasts
}
\description{
PCA reconstruction
}
\examples{
pc_data <- pca(funts(electricity),TVE=0.8)
fit <- .pca_reconstruct(pc_data)
plot_fd(fit)
}
