% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facf.R
\name{.plot_FACF}
\alias{.plot_FACF}
\title{Plot the autocorrelation function of a given FTS}
\usage{
.plot_FACF(rho, SWN, WWN, ...)
}
\arguments{
\item{rho}{Autocorrelation values for each lag of
the functional time series obtained by calling the
function \code{obtain_FACF}.}

\item{SWN}{The upper prediction bound for the strong white noise iid
distribution obtained by calling the function \code{.compute_FACF}.}

\item{WWN}{The upper prediction bound for the weak white noise iid
distribution obtained by calling the function \code{.compute_FACF}.}

\item{...}{Further arguments passed to the \code{plot} function.}
}
\description{
Plot a visual representation of the autocorrelation function of a given
functional time series, including the upper i.i.d. bound.
}
\examples{
# Example 1

N <- 100
v <- seq(from = 0, to = 1, length.out = 10)
sig <- 2
bbridge <- generate_brownian_bridge(N, v, sig)
lag.max <- 15
upper_bound <- 0.95
fACF <- .compute_FACF(X = bbridge, lag.max = lag.max,
                      alpha=upper_bound, figure = FALSE)
.plot_FACF(rho = fACF$acfs,SWN = fACF$SWN_bound, WWN = fACF$WWN_bound)

\donttest{
# Example 2

N <- 200
v <- seq(from = 0, to = 1, length.out = 30)
sig <- 2
bbridge <- generate_brownian_bridge(N, v, sig)
lag.max <- 15
upper_bound <- 0.95
fACF <- .compute_FACF(X = bbridge, lag.max = lag.max,
                      alpha = upper_bound, figure = FALSE)
.plot_FACF(rho = fACF$acfs,SWN = fACF$SWN_bound,WWN = fACF$WWN_bound)
}
}
