% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{.plot_substack}
\alias{.plot_substack}
\title{Plot data as summary stack on 2d plot}
\usage{
.plot_substack(object, CPs = NULL)
}
\arguments{
\item{object}{funts object}

\item{CPs}{CPs, if there are any. Default is NULL.}
}
\value{
Plot (ggplot2) with colored observations / regions
}
\description{
Plot data as summary stack on 2d plot
}
\examples{
# .plot_substack(funts(electricity))
# .plot_substack(funts(electricity), CPs=c(50,100,200))
}
