% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{.summary_nochange}
\alias{.summary_nochange}
\title{Summary for funts Object with no changes}
\usage{
.summary_nochange(object, max.lag = 20, rainbow = c("subset", "full"), ...)
}
\arguments{
\item{object}{funts object}

\item{max.lag}{Max lags to consider for ACF. Default is 20.}

\item{rainbow}{Observations shown in rainbow plot. Defaults to subset,
which summarizes the values rather than showing all observations.}
}
\value{
Plot (ggplot) summarizing the data
}
\description{
Summary for funts Object with no changes
}
\examples{
.summary_nochange(
  generate_brownian_motion(N = 500,
     v=seq(from = 0, to = 1, length.out = 20)))
.summary_nochange(funts(electricity),rainbow='full')
.summary_nochange(generateFAR1(20,300))
}
