% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_change.R
\name{eigen_change}
\alias{eigen_change}
\title{Eigenvalue Changes - Detect eigenvalue changes in the Covariance Operator}
\usage{
eigen_change(
  X,
  d,
  h = 2,
  CPs = NULL,
  test = c("joint", "individual"),
  M = 1000,
  K = bartlett_kernel
)
}
\arguments{
\item{d}{Number of eigenvalues to include in testing.}

\item{h}{The window parameter for the estimation of the long run covariance matrix. The default
value is \code{h=2}.}

\item{M}{Number of monte carlo simulations used to get the critical values. The default value is \code{M=1000}}

\item{fdobj}{A functional data object of class '\code{fd}'}

\item{mean_change}{If \code{TRUE} then the data is centered considering the change in the mean function.}

\item{delta}{Trimming parameter to estimate the covariance function using partial sum estimates.}
}
\value{
\item{\code{pvalue}}{
Approximate p value for testing whether there is a significant change in the desired eigenvalue of the covariance operator
}
\item{\code{change}}{
Estimated change location
}
\item{\code{eval_before}}{
Estimated eigenvalues before the change
}
\item{\code{eval_after}}{
Estimated eigenvalues after the change
}
}
\description{
This function tests and detects changes individual or jointly in the
eigenvalue(s) of the covariance operator.
}
\details{
This function dates and detects changes in the joint eigenvalues
that is defined by \code{d} of the covariance function. The critical
values are approximated via \code{M} Monte Carlo simulations.
}
\examples{
eigen_change(funts(electricity), 2, test='joint')
eigen_change(funts(electricity), 2, test='individual')

\donttest{
bm <- generate_brownian_motion(200,v=seq(0,1,length.out=20))
eigen_change(bm, 3, test='joint')
eigen_change(bm, 3, test='individual')
}
}
\references{
Aue, A, G Rice, and O Sönmez. “Structural Break Analysis
for Spectrum and Trace of Covariance Operators.” Environmetrics
(London, Ont.) 31, no. 1 (2020). \url{https://doi.org/10.1002/env.2617.}
}
