% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{electricity}
\alias{electricity}
\title{Spanish Spot Electricity Data}
\format{
\subsection{\code{electricity}}{

A data frame with 24 rows and 365 columns, representing the hourly (row)
electricity spot values for each day (column) of the year in 2014.
}
}
\source{
<www.omie.es>
}
\usage{
electricity
}
\description{
A fChange ready data set containing the hourly electricity spot prices from
Spain in 2014.
}
\keyword{datasets}
