% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfts.R
\name{[.dfts}
\alias{[.dfts}
\alias{extract}
\alias{[<-.dfts}
\title{Extract or Replace parts of dfts object}
\usage{
\method{[}{dfts}(x, i, j, ...)

\method{[}{dfts}(x, i, j) <- value
}
\arguments{
\item{x}{dfts object}

\item{i, j}{Numerics for elements to extract}

\item{...}{Unused additional parameters}

\item{value}{A suitable replacement value}
}
\value{
dfts object

dfts object
}
\description{
Extract or replace subsets of dfts objects.
}
\examples{
electricity[1:3]
electricity[1:3,]
electricity[1:2,1:4]
electricity[,1:4]
tmp <- dfts(matrix(1:9,3,3))
tmp$data
tmp[1,1] <- 10
tmp$data
}
