% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fChange-package.R
\docType{package}
\name{fChange-package}
\alias{fChange}
\alias{fChange-package}
\title{General Information for fChange}
\description{
Analyze functional data and its change points. Includes functionality to store and process data, summarize and validate assumptions, characterize and perform inference of change points, and provide visualizations. Data is stored as discretely collected observations without requiring the selection of basis functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jrvanderdoes.github.io/fChange/}
}

}
\author{
\strong{Maintainer}: Jeremy VanderDoes \email{jeremy.vanderdoes@gmail.com} (\href{https://orcid.org/0009-0001-9885-3073}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Gregory Rice
  \item Martin Wendler
}

}
\keyword{internal}
