% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{flat_kernel}
\alias{flat_kernel}
\title{Flat Kernel}
\usage{
flat_kernel(x)
}
\arguments{
\item{x}{A numeric value at which to evaluate kernel.
It often indicates current lag divided by window.}
}
\value{
A numeric value indicating value of Kernel at that point
}
\description{
Kernel where \eqn{min(1,max(1.1-|m/h|,0))}
}
\details{
Note, this function is vectorized.
}
\examples{
flat_kernel(-20:20/15)
}
\references{
L. Horvath, P. Kokoszka, G. Rice (2014) "Testing stationarity of functional time series", Journal of
Econometrics, 179(1), 66-82.
}
