% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{flat_top_kernel}
\alias{flat_top_kernel}
\title{Flat-Top Kernel}
\usage{
flat_top_kernel(x)
}
\arguments{
\item{x}{A numeric value at which to evaluate kernel.
It often indicates current lag divided by window.}
}
\value{
A numeric value indicating value of Kernel at that point
}
\description{
Kernel where \eqn{min(1, max(1.1-|x|,0)),|x|\leq 1}. If \eqn{x=0/0} then
the value \eqn{1} is given. Note, values rather than 1.1 have also been
used.
}
\details{
Note, this function is vectorized.
}
\examples{
flat_top_kernel(-20:20/15)
}
\references{
L. Horvath, P. Kokoszka, G. Rice (2014) "Testing stationarity of functional
time series", Journal of Econometrics, 179(1), 66-82.

Politis, D. N. (2003). Adaptive bandwidth choice. Journal of Nonparametric
Statistics, 15(4–5), 517–533. \url{https://doi.org/10.1080/10485250310001604659}

Politis, D. N. (2011). HIGHER-ORDER ACCURATE, POSITIVE SEMIDEFINITE
ESTIMATION OF LARGE-SAMPLE COVARIANCE AND SPECTRAL DENSITY MATRICES.
Econometric Theory, 27(4), 703–744. \url{http://www.jstor.org/stable/27975501}

Horváth, L., & Rice, G. (2024). Change Point Analysis for Time
Series (1st ed. 2024.). Springer Nature Switzerland.
\url{https://doi.org/10.1007/978-3-031-51609-2}
}
