% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{functional_imputation}
\alias{functional_imputation}
\title{Impute data using functional data}
\usage{
functional_imputation(
  data,
  evalPts = 1:nrow(data),
  basis = fda::create.bspline.basis(nbasis = 21, rangeval = c(min(evalPts),
    max(evalPts))),
  ...
)
}
\arguments{
\item{data}{Numeric data.frame with rows for evaluated values and columns
indicating FD}

\item{evalPts}{Numeric vector indicating the evaluated points for each row
of data}

\item{basis}{basis object from fda to fit data}

\item{...}{(Optional) Additional parameters to pass into fda}
}
\value{
Numeric data.frame with rows for evaluated values and columns
indicating FD and no missing values
}
\description{
Impute data using functional data
}
\examples{
data_missing <- data.frame(
  "FD1" = c(1:8, NA, 10) + rnorm(10),
  "FD2" = 21:30 + rnorm(10),
  "FD3" = c(1:3, NA, rep(6, 3), 8, rep(NA, 2)) + rnorm(10)
)
evalPts <- c(1:10)
functional_imputation(data_missing, evalPts)
}
