% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funts.R
\name{funts_group}
\alias{funts_group}
\alias{Math.funts}
\alias{Ops.funts}
\title{Group Generic Functions}
\usage{
\method{Math}{funts}(x, ...)

\method{Ops}{funts}(e1, e2)
}
\arguments{
\item{x, e1, e2}{funts object}

\item{...}{Further arguments passed to the methods}
}
\value{
funts object
}
\description{
Group generic methods defined for things like Math, Ops, and so forth.
}
\examples{
result <- sqrt( funts(electricity) )
result <- funts(electricity) + funts(electricity)
result <- funts(electricity) * funts(electricity)
}
