% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_brownian.R
\name{generate_brownian_bridge}
\alias{generate_brownian_bridge}
\title{Generate a Brownian Bridge Process}
\usage{
generate_brownian_bridge(
  N,
  v = seq(from = 0, to = 1, length.out = 100),
  sd = 1
)
}
\arguments{
\item{N}{Numeric. The number of observations for the generated data.}

\item{v}{Numeric. Discretization points of the curves. By default it is
evenly spaced on \link{0,1}, i.e. \code{seq(from = 0, to = 1, length.out = 100)}.}

\item{sd}{Numeric. Standard deviation of the Brownian Motion process.
The default is \code{1}.}
}
\value{
Functional time series object
}
\description{
Generate a functional time series from an iid Brownian Bridge process.
If \eqn{W(t)} is a Wiener process, the Brownian Bridge is
defined as \eqn{W(t) - tW(1)}.
Each functional observation is discretized in the points
indicated in \code{v}.
}
\examples{
N <- 100
v <- seq(from = 0, to = 1, length.out = 20)
bbridge <- generate_brownian_bridge(N, v, sd=1)
}
