% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_brownian.R
\name{generate_brownian_motion}
\alias{generate_brownian_motion}
\title{Generate a Brownian Motion Process}
\usage{
generate_brownian_motion(
  N,
  v = seq(from = 0, to = 1, length.out = 100),
  sd = 1
)
}
\arguments{
\item{N}{Numeric. The number of observations for the generated data.}

\item{v}{Numeric. Discretization points of the curves. By default it is
evenly spaced on \link{0,1}, i.e. \code{seq(from = 0, to = 1, length.out = 100)}.}

\item{sd}{Numeric. Standard deviation of the Brownian Motion process.
The default is \code{1}.}
}
\value{
Functional time series object
}
\description{
Generate a functional time series according to an iid Brownian Motion process.
Each observation is discretized in the points indicated in \code{v}.
}
\examples{
N <- 100
v <- seq(from = 0, to = 1, length.out = 20)
bmotion <- generate_brownian_motion(N, v, sd = 1)
}
