% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_far.R
\name{generate_far1}
\alias{generate_far1}
\title{Generate FAR(1) Data}
\usage{
generate_far1(resolution, N, d = 1/2, burnin = 1000)
}
\arguments{
\item{resolution}{Numeric for resolution of data}

\item{N}{Numeric for the number of observations}

\item{d}{Numeric (0,Inf) which indicates the dependence on the previous
curve.}

\item{burnin}{Numeric for burn-in before data is computed}
}
\value{
Data.frame for functional observations
}
\description{
Function to generate date according to FAR(1) process.
}
\examples{
res <- generate_far1(24,200)
}
