% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_ou.R
\name{generate_ornstein_uhlenbeck}
\alias{generate_ornstein_uhlenbeck}
\title{Generate Data via Ornstein-Uhlenbeck Process}
\usage{
generate_ornstein_uhlenbeck(N, resolution, rho = 0)
}
\arguments{
\item{N}{Numeric for data length}

\item{resolution}{Numeric for data resolution or specific observed fparam}

\item{rho}{Numeric for amount of dependence}
}
\value{
dfts object of generated OU data
}
\description{
Generate Data via Ornstein-Uhlenbeck Process
}
\examples{
generate_ornstein_uhlenbeck(N=100,resolution=20)
}
