% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_ou.R
\name{generate_ou}
\alias{generate_ou}
\title{Generate Data via OU Process}
\usage{
generate_ou(resolution, N, burnin = 500, rho = 0)
}
\arguments{
\item{resolution}{Numeric for data resolution}

\item{N}{Numeric for data length}

\item{burnin}{Numeric for number of burnin. Default is 500.}

\item{rho}{Numeric for amount of dependence}
}
\value{
funts object of generated OU data
}
\description{
Generate Data via OU Process
}
\examples{
generate_ou(20,100)
}
