% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_far.R
\name{generateFAR1}
\alias{generateFAR1}
\title{Title}
\usage{
generateFAR1(resolution, N, d = 1/2, burnin = 1000)
}
\arguments{
\item{d}{Numeric (0,Inf) which indicates the dependence on the previous
curve.}

\item{burnin}{}
}
\description{
Title
}
\examples{
res <- .generateFAR1(24,200)
}
