% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facf_functions1.R, R/tests_wn_ind_functions.R
\name{iid_covariance}
\alias{iid_covariance}
\title{Partial Covariance under iid}
\usage{
iid_covariance(f_data)

iid_covariance(f_data)
}
\arguments{
\item{f_data}{the functional data matrix with observed functions in the columns}
}
\value{
A 2-dimensional matrix containing one of the two independent sums in the computation of the
covariance.

A 2-dimensional matrix containing one of the two independent sums in the computation of the
covariance.
}
\description{
Compute part of the covariance under a strong white noise assumption

\code{iid_covariance} A helper function used to compute one of the two independent sum terms in the
computation of the approximate covariance of the functional data under a strong white noise assumption.
}
\details{
\code{iid_covariance} A helper function used to compute one of the two independent sum terms in the
`computation of the approximate covariance of the functional data under a strong white noise assumption.
}
