% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_wn_ind.R
\name{imhof_test}
\alias{imhof_test}
\title{Imhof Test}
\usage{
imhof_test(data, lag)
}
\arguments{
\item{data}{funts object or data that can be easily converted}

\item{lag}{the lag for which to compute the imhof test}
}
\value{
A list containing the SVD of tensor c^hat_i_j(t,s,u,v) and the p-value computing the
probability that the observed value of the statistic Q_h is larger than the 1-alpha quantile
of the quadratic form in normal variables.
}
\description{
Returns the the SVD of the tensor c^hat_i_j(t,s,u,v) and the p-value computing
the probability that the observed value of the statistic Q_h is larger than the 1-alpha
quantile of the quadratic form in normal variables described in (15)
}
\examples{
imhof_test(funts(electricity),1)
}
