% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funts.R
\name{is.funts}
\alias{is.funts}
\title{Check data to see if its funts}
\usage{
is.funts(x)
}
\arguments{
\item{x}{Data to examine class}
}
\value{
Boolean indicating if \code{x} is a funts object or not
}
\description{
Check data to see if its funts
}
\examples{
# is.funts(electricity)
# is.funts(funts(electricity))
}
