% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_eigen_change.R
\name{joint_eigen_change}
\alias{joint_eigen_change}
\title{Detecting Changes Jointly in the Eigenvalues of the Covariance Operator of the Functional Data}
\usage{
joint_eigen_change(
  X,
  d,
  h = 2,
  CPs = NULL,
  delta = 0.1,
  M = 1000,
  K = bartlett_kernel
)
}
\arguments{
\item{d}{Number of eigenvalues to include in testing.}

\item{h}{The window parameter for the estimation of the long run covariance matrix. The default
value is \code{h=2}.}

\item{delta}{Trimming parameter to estimate the covariance function using partial sum estimates.}

\item{M}{Number of monte carlo simulations used to get the critical values. The default value is \code{M=1000}}

\item{fdobj}{A functional data object of class '\code{fd}'}

\item{mean_change}{If \code{TRUE} then the data is centered considering the change in the mean function.}
}
\value{
\item{\code{pvalue}}{
Approximate p value for testing whether there is a significant change in the desired eigenvalue of the covariance operator
}
\item{\code{change}}{
Estimated change location
}
\item{\code{eval_before}}{
Estimated eigenvalues before the change
}
\item{\code{eval_after}}{
Estimated eigenvalues after the change
}
}
\description{
This function tests and detects changes jointly in the eigenvalue of
the covariance operator.
}
\details{
This function dates and detects changes in the joint eigenvalues
that is defined by \code{d} of the covariance function. The critical
values are approximated via \code{M} Monte Carlo simulations.
}
\examples{
joint_eigen_change(generate_brownian_motion(500), 1)
joint_eigen_change(funts(electricity), 2)
}
