% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kernels}
\alias{kernels}
\alias{truncated_kernel}
\alias{bartlett_kernel}
\alias{parzen_kernel}
\alias{tukey_hanning_kernel}
\alias{quadratic_spectral_kernel}
\alias{daniell_kernel}
\alias{flat_top_kernel}
\title{Kernel Functions}
\usage{
truncated_kernel(x)

bartlett_kernel(x)

parzen_kernel(x)

tukey_hanning_kernel(x)

quadratic_spectral_kernel(x)

daniell_kernel(x)

flat_top_kernel(x)
}
\arguments{
\item{x}{A numeric value at which to evaluate kernel.
It often indicates current lag divided by window.}
}
\value{
Values from a given lag in the kernel
}
\description{
There are an assortment of (vectorized) kernel functions located in the package.
}
\examples{
truncated_kernel(-20:20/15)
bartlett_kernel(-20:20/15)
parzen_kernel(-20:20/15)
tukey_hanning_kernel(-20:20/15)
quadratic_spectral_kernel(-20:20/15)
daniell_kernel(-20:20/15)
flat_top_kernel(-20:20/15)
}
\references{
Horvath, L., & Rice, G. (2024). Change Point Analysis for Time
Series (1st ed. 2024.). Springer Nature Switzerland.
\url{https://doi.org/10.1007/978-3-031-51609-2}

L. Horvath, P. Kokoszka, G. Rice (2014) "Testing stationarity of functional
time series", Journal of Econometrics, 179(1), 66-82.

Politis, D. N. (2003). Adaptive bandwidth choice. Journal of Nonparametric
Statistics, 15(4-5), 517-533. \url{https://doi.org/10.1080/10485250310001604659}

Politis, D. N. (2011). HIGHER-ORDER ACCURATE, POSITIVE SEMIDEFINITE
ESTIMATION OF LARGE-SAMPLE COVARIANCE AND SPECTRAL DENSITY MATRICES.
Econometric Theory, 27(4), 703-744. \url{http://www.jstor.org/stable/27975501}
}
