% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpss_test.R
\name{kpss_test}
\alias{kpss_test}
\title{Functional KPSS Test}
\usage{
kpss_test(
  X,
  method = c("MC", "block", "sliding"),
  M = 1000,
  h = 3,
  TVE = 1,
  replace = TRUE
)
}
\arguments{
\item{X}{A dfts object or data which can be automatically converted to that
format. See \code{\link[=dfts]{dfts()}}.}

\item{method}{String (default MC) for the method: Monte Carlo simulation
(\code{MC}), block resampling (\code{block}), and sliding window
permutation (\code{sliding})}

\item{M}{Number of simulations (default 1000) to estimate theoretical distribution}

\item{h}{Numeric (default 3) for the block size when using a resample test}

\item{TVE}{Numeric (default 1) for selecting the number of principle
components / eigenvalues}

\item{replace}{Boolean (default TRUE) to indicate if blocks should be
selected with replacement when using a resample test}
}
\value{
List with statistic, p-value and theoretical simulations, respectively
}
\description{
Compute the KPSS statistic for functional data.
}
\examples{
kpss_test(generate_brownian_motion(100,v=seq(0,1,length.out=20)))
kpss_test(generate_brownian_motion(100,v=seq(0,1,length.out=20)),
             method="block")
kpss_test(generate_brownian_motion(100,v=seq(0,1,length.out=20)),
             method='sliding')
}
\references{
Chen, Y., & Pun, C. S. (2019). A bootstrap-based KPSS test for
functional time series. Journal of Multivariate Analysis, 174, 104535.

Kokoszka, P., & Young, G. (2016). KPSS test for functional time
series. Statistics, 50(5), 957-973.
}
