% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_change.R
\name{kweights}
\alias{kweights}
\title{Kernels for applying weights}
\usage{
kweights(
  x,
  kernel = c("Truncated", "Bartlett", "Parzen", "Tukey-Hanning", "Quadratic Spectral"),
  normalize = FALSE
)
}
\arguments{
\item{x}{XXXXXX}

\item{kernel}{XXXXXX}

\item{normalize}{XXXXXX}
}
\value{
XXXXX
}
\description{
Kernels for applying weights
}
\examples{
# XXXXX
}
