% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfts.R
\name{lag.dfts}
\alias{lag.dfts}
\title{Lag dfts}
\usage{
\method{lag}{dfts}(x, k = 1, ...)
}
\arguments{
\item{x}{A dfts object or data which can be automatically converted to that
format. See \code{\link[=dfts]{dfts()}}.}

\item{k}{integer indicating the number of lags for the data}

\item{...}{Unused additional parameters}
}
\value{
A dfts object
}
\description{
Lag dfts
}
\examples{
result <- lag(electricity)
}
