% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funts.R
\name{lag.funts}
\alias{lag.funts}
\title{Lag funts}
\usage{
\method{lag}{funts}(x, k, ...)
}
\arguments{
\item{x}{funts object}

\item{k}{integer indicating the number of lags for the data}

\item{...}{Unused additional parameters}
}
\value{
A funts object
}
\description{
Lag funts
}
\examples{
lag.funts(funts(electricity))
}
