% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_run_cov.R
\name{long_run_covariance}
\alias{long_run_covariance}
\title{Estimate Long-run Covariance Kernel}
\usage{
long_run_covariance(X, h = 0, K = bartlett_kernel)
}
\arguments{
\item{X}{A dfts object or data which can be automatically converted to that
format. See \code{\link[=dfts]{dfts()}}.}

\item{h}{The window parameter parameter for the estimation of the long run
covariance kernel. The default value is \code{h=0}, i.e., it assumes iid data}

\item{K}{(Optional) Function indicating the Kernel to use if h>0}
}
\value{
Data.frame of numerics with dim of ncol(data) x ncol(data), that is
symmetric.
}
\description{
This (internal) function estimates the long-run covariance kernel. That is,
\eqn{C_{\epsilon}(t,t') = \sum_{l=-\inf}^{\inf} \text{Cov}(\epsilon_0(t),
    \epsilon_l(t'))} with error sequence \eqn{(\epsilon_i : i \in \mathbb{Z})}.
}
\details{
This is an internal function, see use in .change_mean.
}
\examples{
result <- long_run_covariance(electricity,2)
}
