% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_change.R
\name{long_run_covariance_4tensor}
\alias{long_run_covariance_4tensor}
\title{Long-run covariance estimator using tensor operator}
\usage{
long_run_covariance_4tensor(dat)
}
\arguments{
\item{dat}{An array with dimension (grid_point,grid_point,N)}
}
\value{
XXXXX
}
\description{
Long-run covariance estimator using tensor operator
}
\examples{
# XXXXX
}
