% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funts.R
\name{Math.funts}
\alias{Math.funts}
\title{Math Group for funts}
\usage{
\method{Math}{funts}(x, ...)
}
\arguments{
\item{x}{funts object}

\item{...}{Additional parameters for functions in math group}
}
\value{
funts object with applied functions
}
\description{
Math Group for funts
}
\examples{
# sqrt( funts(electricity) )
}
