% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funts.R
\name{max.funts}
\alias{max.funts}
\title{Max funts}
\usage{
\method{max}{funts}(x, ...)
}
\arguments{
\item{x}{funts object}

\item{...}{Unused additional parameters}
}
\value{
A funts object
}
\description{
Get the observation(s) with the max average value.
}
\details{
TODO:: Add a pw option
}
\examples{
# max(funts(electricity))
}
