% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facf_functions.R, R/tests_wn_ind_functions.R
\name{MCint_eta_approx_i_j}
\alias{MCint_eta_approx_i_j}
\title{Monte Carlo Integration for eta_ij}
\usage{
MCint_eta_approx_i_j(f_data, i, j, M = NULL, low_disc = FALSE)

MCint_eta_approx_i_j(f_data, i, j, M = NULL, low_disc = FALSE)
}
\arguments{
\item{f_data}{the functional data matrix with functions in columns}

\item{i, j}{the indices i,j in 1:T that we are computing eta^hat_i_j for}

\item{M}{number of vectors (v1, v2, v3, v4) to sample uniformly from U_J X U_J X U_J X U_J}

\item{low_disc}{boolean value specifying whether or not to use low-discrepancy sampling
for the Monte-Carlo method (only Sobol Sampling is currently supported)}
}
\value{
scalar value of eta^_hat_i_j computed using the MCint method.

scalar value of eta^_hat_i_j computed using the MCint method.
}
\description{
Computes an approximation of eta_i_j (defined under (15)) using the second
Monte Carlo integration method "MCint" defined on page 8.

Computes an approximation of eta_i_j (defined under (15)) using the second
Monte Carlo integration method "MCint" defined on page 8.
}
