% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funts.R
\name{mean.funts}
\alias{mean.funts}
\title{Mean for funts}
\usage{
\method{mean}{funts}(x, ...)
}
\arguments{
\item{x}{funts object}

\item{...}{Unused additional parameters}
}
\value{
Numeric for the mean at each resolution
}
\description{
Mean for funts
}
\examples{
# mean(funts(electricity))
}
