% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_change.R
\name{mean_change}
\alias{mean_change}
\title{Fully Functional Mean Change Point Analysis}
\usage{
mean_change(
  data,
  M = 1000,
  h = 0,
  K = bartlett_kernel,
  alpha = 0.05,
  inc.pval = FALSE,
  ...
)
}
\arguments{
\item{data}{Numeric data.frame with evaled points on rows and fd objects in columns}

\item{M}{(Optional) Number of Monte Carlo simulations used to get the critical
values. The default value is \code{M=1000}}

\item{h}{(Optional) The window parameter parameter for the estimation of the
long run covariance kernel. The default value is \code{h=0}, i.e., it
assumes iid data}

\item{K}{(Optional) Function indicating the Kernel to use if h>0}

\item{alpha}{(Optional) Numeric value indicating significance. Defaults to 0.05}

\item{inc.pval}{(Optional) Boolean to indicate if pval should also be returned}

\item{...}{Unused, just for use in other functions}
}
\value{
If inc.pval is false, Numeric of CP location or NA, otherwise
location or NA and the p-value
}
\description{
This function tests whether there is a significant change in the mean
function of the functional data, and it will give an estimate for the
location of the change. The procedure is based on the standard L-2 norm
and hence does not depend on any dimension reduction technique such as
fPCA.
}
\examples{
mean_change(electricity,M=250)
}
\references{
Aue A., Rice G., Sonmez O. (2017+), \emph{Detecting and dating structural breaks in
functional data without dimension reduction} (https://arxiv.org/pdf/1511.04020.pdf)
}
