% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facf_functions.R, R/tests_wn_ind_functions.R
\name{mean_hat_Q_h}
\alias{mean_hat_Q_h}
\title{FACF - Welch Approximation Mean Hat}
\usage{
mean_hat_Q_h(f_data, lags)

mean_hat_Q_h(f_data, lags)
}
\arguments{
\item{f_data}{the functional data matrix with functions in columns}

\item{lags}{specifies the lag use in the test statistic Q_h (lag = h in paper)}
}
\value{
scalar approximation of the mean of the test statistic Q_h.

scalar approximation of the mean of the test statistic Q_h.
}
\description{
Computes the approximation of the mean defined in (15) which is used in the Welch-
Satterthwaite approximation as mean of the chi-squared random variable approximating Q_h.

Computes the approximation of the mean defined in (15) which is used in the Welch-
Satterthwaite approximation as mean of the chi-squared random variable approximating Q_h.
}
