% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facf_functions1.R, R/tests_wn_ind_functions.R
\name{mean_hat_Q_h_iid}
\alias{mean_hat_Q_h_iid}
\title{Test Statistic - Mean Hat Q_h_iid}
\usage{
mean_hat_Q_h_iid(f_data)

mean_hat_Q_h_iid(f_data)
}
\arguments{
\item{f_data}{the functional data matrix with functions in columns}
}
\value{
scalar approximation of the mean of the test statistic Q_h under a strong white noise
assumption.

scalar approximation of the mean of the test statistic Q_h under a strong white noise
assumption.
}
\description{
Computes the approximation of the mean defined in (15) which is used in the
Welch-Satterthwaite approximation under the assumption that the functional data follows a
strong white noise.

Computes the approximation of the mean defined in (15) which is used in the
Welch-Satterthwaite approximation under the assumption that the functional data follows a
strong white noise.
}
