% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_wn_ind_functions.R
\name{mean_hat_V_K}
\alias{mean_hat_V_K}
\title{Mean Approximation - mean_hat_V_K}
\usage{
mean_hat_V_K(f_data, K)
}
\arguments{
\item{f_data}{the functional data matrix with functions in columns}

\item{K}{specifies the range of lags 1:K for for the test statistic V_K}
}
\value{
scalar approximation of the mean of the test statistic V_K.
}
\description{
Computes the approximation of the mean defined in (15) which is used in the Welch-
Satterthwaite approximation as mean of the chi-squared random variable approximating V_K.
}
