% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_wn_ind_functions.R
\name{mean_hat_V_K_iid}
\alias{mean_hat_V_K_iid}
\title{Mean Approximation - iid mean_hat_V_K}
\usage{
mean_hat_V_K_iid(f_data, K)
}
\arguments{
\item{f_data}{the functional data matrix with functions in columns}

\item{K}{specifies the range of lags 1:K for for the test statistic V_K}
}
\value{
scalar approximation of the mean of the test statistic V_K under a strong white noise
assumption.
}
\description{
Computes the approximation of the mean defined in (15) which is used in the
Welch-Satterthwaite approximation under the assumption that the functional data follows a
strong white noise.
}
