% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_pca_change.R
\name{mean_pca_change}
\alias{mean_pca_change}
\title{Change Point Analysis - Dimension Reduction}
\usage{
mean_pca_change(X, TVE = 0.95, M = 1000, h = 0, K = bartlett_kernel)
}
\arguments{
\item{M}{Number of monte carlo simulations to get the critical values. The default value is \code{M=1000}}

\item{h}{The window parameter for the estimation of the long run covariance kernel. The default
value is \code{h=0}, i.e., it assumes iid data}

\item{fdobj}{A functional data object of class '\code{fd}'}

\item{d}{Number of principle components}

\item{plot}{If \code{TRUE} plot of the functional data before and after the estimated change and plot of the
estimated change function is given}

\item{...}{Further arguments to pass}
}
\value{
\item{\code{pvalue}}{
An approximate p value for testing whetehr there is a significant change in the mean function
}
\item{\code{change}}{
Estimated change location
}
\item{\code{DataBefore}}{
Data before the estimated change
}
\item{\code{DataAfter}}{
Data after the estimated change
}
\item{\code{MeanBefore}}{
Mean function before the estimated change
}
\item{\code{MeanAfter}}{
Mean function after the estimated change
}
\item{\code{change_fun}}{
Estimated change function
}
}
\description{
This function tests whether there is a significant change in the mean
function of functional data, and it gives an estimate of the location
of the change. The procedure will reduce the dimension of the
functional data using functional principal component analysis and will
use the leading principal curves which explain \code{TVE} total variance
to carry out the change point analysis.
}
\details{
This functions performs structural break analysis for the functional
data using an fPCA based initial dimension reduction. It is recommended
that the dimension of the subspace, \code{d}, that the functional
observations are projected onto should be selected based on TVE using
\code{\link{pick_dim}}.
}
\examples{
res <- mean_pca_change(generate_brownian_bridge(200,seq(0,1,length.out=10)))
res1 <- mean_pca_change(generate_ou(20,200))
res2 <- mean_pca_change(funts(electricity))
}
\references{
Berkes, I., Gabrys, R.,Horvath, L. & P. Kokoszka (2009).,
\emph{Detecting changes in the mean of functional observations}
Journal of the Royal Statistical Society, Series B 71, 927–946

Aue, A., Gabrys, R.,Horvath, L. & P. Kokoszka (2009).,
\emph{Estimation of a change-point in the mean function of functional data}
Journal of Multivariate Analysis 100, 2254–2269.
}
\seealso{
\code{\link{change_FF}}
}
