% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions_existing.R
\name{median.funts}
\alias{median.funts}
\title{Median for funts}
\usage{
\method{median}{funts}(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{funts object}

\item{...}{Unused additional parameters}
}
\value{
Numeric for the median at each resolution
}
\description{
Median for funts
}
\examples{
# median(funts(electricity))
}
