% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funts.R
\name{min.funts}
\alias{min.funts}
\title{Min funts}
\usage{
\method{min}{funts}(x, ...)
}
\arguments{
\item{x}{funts object}

\item{...}{Unused additional parameters}
}
\value{
A funts object
}
\description{
Get the observation(s) with the min average value
}
\details{
TODO:: Add a pw option
}
\examples{
# min(funts(electricity))
}
