% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelation.R
\name{obtain_autocovariance}
\alias{obtain_autocovariance}
\title{Estimate the autocovariance function of the series}
\usage{
obtain_autocovariance(X, nlags)
}
\arguments{
\item{X}{funts object}

\item{nlags}{Number of lagged covariance operators of the functional
time series that will be used to estimate the autocorrelation function.}
}
\value{
Return a list with the lagged autocovariance functions
estimated from the data. Each function is given by a \eqn{(m x m)}
matrix, where \eqn{m} is the number of points observed in each curve.
}
\description{
Obtain the empirical autocovariance function for lags
\eqn{= 0,...,}\code{nlags} of the functional time series. Given
\eqn{Y_{1},...,Y_{T}} a functional time series, the sample autocovariance
functions \eqn{\hat{C}_{h}(u,v)} are given by:
\deqn{\hat{C}_{h}(u,v) =  \frac{1}{T} \sum_{i=1}^{T-h}(Y_{i}(u) - \overline{Y}_{T}(u))(Y_{i+h}(v) - \overline{Y}_{T}(v))}
where
\eqn{ \overline{Y}_{T}(u) = \frac{1}{T} \sum_{i = 1}^{T} Y_{i}(t)}
denotes the sample mean function.
}
\examples{
v <- seq(0,1,length.out=20)
lagged_autocov <- obtain_autocovariance(
  X = generate_brownian_bridge(100,v=v),
  nlags = 1)
image(x = v, y = v, z = lagged_autocov$Lag0)
}
