% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funts.R
\name{Ops.funts}
\alias{Ops.funts}
\title{Ops Group for funts}
\usage{
\method{Ops}{funts}(e1, e2)
}
\arguments{
\item{e1}{funts object}

\item{e2}{funts object}
}
\value{
funts object from the two elements
}
\description{
TODO: Figure out for non-intersections.
}
\examples{
funts(electricity) + funts(electricity)
funts(electricity) * funts(electricity)
}
