% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{pacf.funts}
\alias{pacf.funts}
\title{PACF for Functional Data}
\usage{
\method{pacf}{funts}(x, ...)
}
\arguments{
\item{...}{Additional parameters to function based on data}
}
\value{
Functional PACF values
}
\description{
PACF for Functional Data
}
\examples{
pacf(funts(electricity))
}
