% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{parzen_kernel}
\alias{parzen_kernel}
\title{Parzen Kernel}
\usage{
parzen_kernel(x)
}
\arguments{
\item{x}{A numeric value at which to evaluate kernel.
It often indicates current lag divided by window.}
}
\value{
A numeric value indicating value of Kernel at that point
}
\description{
Kernel where \eqn{1 - 6 * x^2 + 6 * |x|^3, |x|<=0.5},
\eqn{ 2 * (1 - |x|)^3, 0.5<|x|<1}, and \eqn{0, |x|>1}.
}
\details{
Note, this function is vectorized.
}
\examples{
parzen_kernel(-20:20/15)
}
