% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{pca.funts}
\alias{pca.funts}
\title{Principal component analysis}
\usage{
\method{pca}{funts}(x, TVE = 1, ...)
}
\arguments{
\item{TVE}{Numeric in \link{0.1} for the total variance explained. Can use this to
select only the required components}

\item{...}{Additional parameters for \code{prcomp}.}
}
\value{
Principal component data
}
\description{
Principal component analysis
}
\examples{
pca(funts(electricity))
}
