% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{pca}
\alias{pca}
\alias{pca.default}
\title{Generic pca of Data}
\usage{
pca(object, ...)

\method{pca}{default}(object, ...)
}
\arguments{
\item{object}{Data to compute principal component analysis on}

\item{...}{Additional information for method}
}
\value{
PCA data

PCA object
}
\description{
Build on \code{prcomp}.
}
\examples{
pca(1:10)
}
