% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_examination.R
\name{pca_components}
\alias{pca_components}
\title{Functional PCA Components}
\usage{
pca_components(pca, components = 1:length(pca$sdev))
}
\arguments{
\item{pca}{PCA object from \code{pca()}}

\item{components}{Numeric for the components of interest. Can be a single
numeric to examine that component or multiple to examine the combined result}
}
\value{
A dfts object of the pca component(s)
}
\description{
Evaluates the given pca components as a function.
}
\examples{
tmp <- pca(electricity, TVE = 0.1)
pca_components(tmp, components = 1)
}
