% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_examination.R
\name{pca_examination}
\alias{pca_examination}
\title{Principal Component Exploration}
\usage{
pca_examination(X, TVE = 0.95, d.max = 3)
}
\arguments{
\item{X}{funts object}

\item{TVE}{Numeric in [0,1]. Total variance explained to determine the
number of components to examine}

\item{d.max}{Numeric. Max number of components to investigate (if TVE
suggests too many).}
}
\value{
List with the following elements:
\itemize{
\item Figures: List with figures on all components, summaries, reconstructed
values, and residuals
\item reconstruction: Reconstructed value from pcs
\item residuals: Difference between true and reconstruction
}
}
\description{
Principal Component Exploration
}
\examples{
results <- pca_examination(funts(electricity))
}
