% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_examination.R
\name{pca_fit}
\alias{pca_fit}
\title{Forecast using principal component analysis}
\usage{
pca_fit(X, TVE = 0.95, model = c("ets", "arima"), ...)
}
\arguments{
\item{X}{funts object}

\item{TVE}{Numeric for the total variance explained to select number
of components in PCA}

\item{model}{String to indicate method to model components, either
"ets" or "arima".}

\item{...}{Additional parameters to pass into pca}
}
\value{
List with the following elements:
\itemize{
\item fit: funts object for fit
\item errors: funts object for errors from fit
\item parameters: list with fit parameters like pcs, TVE, and model
}
}
\description{
TODO:: Add n to forecast ahead
}
\examples{
pca_fit(funts(electricity))
pca_fit(generate_brownian_motion(100))
}
