% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_examination.R
\name{pcaExploration}
\alias{pcaExploration}
\title{Principal Component Exploration}
\usage{
pcaExploration(X, order = 3, ...)
}
\arguments{
\item{X}{funts object}

\item{order}{Number of components to investigate}

\item{...}{Unused additional inputs}
}
\value{
List with the following elements:
\itemize{
\item Figures: List with figures on all components, summaries, reconstructed
values, and residuals
\item reconstruction: Reconstructed value from pcs
\item residuals: Difference between true and reconstruction
}
}
\description{
Principal Component Exploration
}
\examples{
results <- pcaExploration(funts(electricity))
}
