% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{A value or the tidyr placeholder.}

\item{rhs}{A function call using the tidyr semantics.}
}
\value{
The result of calling \code{rhs(lhs)}.
}
\description{
See \code{tidyr::\link[tidyr:pipe]{\%>\%}} for details.
}
\keyword{internal}
