% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.funts}
\alias{plot.funts}
\title{Plot Funts}
\usage{
\method{plot}{funts}(
  x,
  CPs = NULL,
  type = c("spaghetti", "highdim", "rainbow", "banded", "acf", "pacf", "summary", "qq",
    "distribution", "change", "interval", "surface"),
  plot_title = x$name,
  val_axis_title = NULL,
  res_axis_title = NULL,
  FD_axis_title = NULL,
  eye = NULL,
  aspectratio = NULL,
  showticklabels = TRUE,
  lag.max = 20,
  d.max = 2,
  alpha = 0.05,
  TVE = 0.95,
  distribution = c("norm"),
  method = c("Welch", "MC", "Imhof"),
  legend = TRUE,
  highlight_changes = TRUE,
  intervals = confidence_interval(x, CPs),
  int.gradual = TRUE,
  ...
)
}
\arguments{
\item{x}{Funts object or data easily convertible. See \code{\link[=funts]{funts()}}.}

\item{CPs}{Vector of change points. Can be NULL.}

\item{type}{Choice of plotting method. Options include: 'spaghetti', 'highdim',
'rainbow','banded','acf', 'pacf', 'summary', 'qq', 'distribution', 'change',
'interval', and'surface'.}

\item{plot_title}{Title to include on the return plot}

\item{val_axis_title, res_axis_title, FD_axis_title}{Title for the axis giving
the values (val), the resolution of the intraobs (res), and the functional
observations (FD)}

\item{eye, aspectratio}{Angle (eye) and ratio (aspectratio) to view 3d plots}

\item{showticklabels}{Boolean if the tick marks should be shown}

\item{lag.max}{Max number of lags to consider for ACF/PACF and summary plots}

\item{d.max}{Max number of dimensions for qq/distribution and summary plots}

\item{alpha}{Significance level to be used in various plots. Value in [0,1]}

\item{TVE}{Total variance explained used in qq/distribution plots. Value in [0,1]}

\item{distribution}{String of the distribution to compare against in
distribution plot. The string can be anything such that there is a
rdistribution and ddistribution function available. For example "exp",
"gamma". Additional parameters can be passed using ...}

\item{method}{Method for computing ACF/PACF. Options include 'Welch', 'MC',
and 'Imhof'.}

\item{legend}{Boolean if legend should be given in qq/distribution plots}

\item{highlight_changes, int.gradual}{Boolean if changes should be highlighted
or confidence interval be solid gray or gradual colors in change plot}

\item{intervals}{Information on confidence intervals of changes for change
plot. See \code{\link[=confidence_interval]{confidence_interval()}}.}

\item{...}{Details for plotting in acf/pacf, summary, or distribution function}
}
\value{
Plot of varying types
}
\description{
A function to plot funts objects in many ways.
}
\examples{
plt <- plot(funts(electricity))
plt <- plot(var(funts(electricity)), type='surface')
}
\references{
John Fox, & Sanford Weisberg (2019). An R Companion to Applied
Regression. Sage.
}
