% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fd.R
\name{plot_fd}
\alias{plot_fd}
\title{Plot functional data}
\usage{
plot_fd(data, eval_points = NULL, plot_title = NULL)
}
\arguments{
\item{data}{An fd object or a data.frame of evaluated fd objects
(the columns bieng lines and the rows the evaluated points)}

\item{eval_points}{An optional vector containing the points at which the
evaluated data was evaluated at}

\item{plot_title}{Optional string to title the plot}
}
\value{
A plotly plot
}
\description{
\code{plot_fd} plots functional data either in an fd object or evaluated at certain points
}
\examples{
# Generate Data with a mean change point
evalPts <- seq(0,1,0.05)
data_KL <- generate_data_fd(ns = c(100,100),
                               eigsList = list(c(3,2,1,0.5),
                                               c(3,2,1,0.5)),
                               basesList = list(fda::create.bspline.basis(nbasis=4, norder=4),
                                                fda::create.bspline.basis(nbasis=4, norder=4)),
                               meansList = c(-1,1),
                               distsArray = c('Normal','Normal'),
                               evals = evalPts,
                               kappasArray = c(0,0))
plot_fd(data=data_KL,eval_points=evalPts)

# Plot an FD object. Note differences are due to plotting the smoothed data
data_fd <- fda::Data2fd(argvals=evalPts, y=as.matrix(data_KL),
                   basisobj=fda::create.bspline.basis(rangeval=range(evalPts)))
plot_fd(data=data_fd)
}
