% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfts.R
\name{print.dfts}
\alias{print.dfts}
\title{Printing dfts objects}
\usage{
\method{print}{dfts}(x, ...)
}
\arguments{
\item{x}{dfts data}

\item{...}{unused parameter}
}
\description{
Basic formatting to print a dfts object
}
\examples{
electricity
}
