% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_examination.R
\name{projection_model}
\alias{projection_model}
\title{Forecast using principal component analysis}
\usage{
projection_model(
  X,
  TVE = 0.95,
  model = c("ets", "arima"),
  n.ahead = 0,
  alpha = 0.05,
  check.cp = TRUE,
  ...
)
}
\arguments{
\item{X}{A dfts object or data which can be automatically converted to that
format. See \code{\link[=dfts]{dfts()}}.}

\item{TVE}{Numeric for the total variance explained to select number
of components in PCA}

\item{model}{String to indicate method to model components, either
"ets" or "arima".}

\item{n.ahead}{Number of observations to forecast}

\item{alpha}{Significance in [0,1] for intervals when forecasting}

\item{check.cp}{Boolean which indicates if the errors should be checked for
change points to change forecasts and plots}

\item{...}{Additional information to pass into pca, change (if check.cp=TRUE),
and plot}
}
\value{
List with the following elements:
\itemize{
\item fit: dfts object for fit
\item errors: dfts object for errors from fit
\item parameters: list with fit parameters like pcs, TVE, and model
}
}
\description{
Forecast using principal component analysis
}
\examples{
result <- projection_model(dfts(electricity$data[,50:100]),
 n.ahead=1, TVE=0.1, check.cp=FALSE)
}
\references{
Hyndman, R. J., & Shahid Ullah, M. (2007). Robust forecasting
of mortality and fertility rates: A functional data approach. Computational
Statistics & Data Analysis, 51(10), 4942-4956.
https://doi.org/10.1016/j.csda.2006.07.028
}
