% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facf_functions.R, R/tests_wn_ind_functions.R
\name{Q_WS_quantile}
\alias{Q_WS_quantile}
\title{FACF - Welch Approximation}
\usage{
Q_WS_quantile(f_data, lag, alpha = 0.05, M = NULL, low_disc = FALSE)

Q_WS_quantile(f_data, lag, alpha = 0.05, M = NULL, low_disc = FALSE)
}
\arguments{
\item{f_data}{the functional data matrix with functions in columns}

\item{lag}{specifies the lag used for the test statistic Q_h}

\item{alpha}{the significance level to be used in the hypothesis test}

\item{M}{optional argument specifying the sampling size in the related Monte Carlo method}

\item{low_disc}{Boolean value specifying whether or not to use low-discrepancy sampling
for the Monte-Carlo method (only Sobol Sampling is currently supported)}
}
\value{
scalar value of the 1-alpha quantile of the beta * chi-square distribution with nu
degrees of freedom (which approximates Q_h).

scalar value of the 1-alpha quantile of the beta * chi-square distribution with nu
degrees of freedom (which approximates Q_h).
}
\description{
Computes the 1-alpha quantile of the beta * chi-squared distribution with nu
degrees of freedom, where beta and nu are obtained from a Welch-Satterthwaite approximation
of the test statistic Q_h. This quantile is used to conduct an approximate size alpha test
of the hypothesis H_0_h.

Computes the 1-alpha quantile of the beta * chi-squared distribution with nu
degrees of freedom, where beta and nu are obtained from a Welch-Satterthwaite approximation
of the test statistic Q_h. This quantile is used to conduct an approximate size alpha test
of the hypothesis H_0_h.
}
