% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portmanteau_acf_functions.R
\name{Q_WS_quantile_iid}
\alias{Q_WS_quantile_iid}
\title{Test Statistic - Welch Approximation for iid Q_WS_hyp_test}
\usage{
Q_WS_quantile_iid(f_data, alpha = 0.05)
}
\arguments{
\item{f_data}{the functional data matrix with functions in columns}

\item{alpha}{the significance level to be used in the hypothesis test}
}
\value{
scalar value of the 1-alpha quantile of the beta * chi-square distribution with nu
degrees of freedom (which approximates Q_h) (computed under a strong white noise
assumption).

@noRd
@keywords internal
}
\description{
Computes the size alpha test of the hypothesis H_0_h using the WS
Approximation under the assumption that the data follows a strong white noise.
}
