% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq.R
\name{qqplot.dfts}
\alias{qqplot.dfts}
\title{Generic Function to Compute QQ plot for dfts}
\usage{
\method{qqplot}{dfts}(
  x,
  TVE = 0.95,
  d.max = 3,
  alpha = 0.05,
  changes = NULL,
  legend = FALSE,
  ...
)
}
\arguments{
\item{x}{A dfts object or data which can be automatically converted to that
format. See \code{\link[=dfts]{dfts()}}.}

\item{TVE}{Numeric in [0,1] giving the total variance explained for selecting
the number of principal components.}

\item{d.max}{Max number of principal components}

\item{alpha}{Significance level, alpha in [0,1]}

\item{changes}{Vector of change points}

\item{legend}{Boolean indicating if legend should be shown}

\item{...}{Unused}
}
\value{
ggplot2 for QQ plot
}
\description{
Creates normal QQ plots on the principal components of functional data
}
\examples{
result <- qqplot(electricity)
}
