% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{quadratic_spectral_kernel}
\alias{quadratic_spectral_kernel}
\title{Quadratic Spectral Kernel}
\usage{
quadratic_spectral_kernel(x)
}
\arguments{
\item{x}{A numeric value at which to evaluate kernel.
It often indicates current lag divided by window.}
}
\value{
A numeric value indicating value of Kernel at that point
}
\description{
Kernel where \eqn{\frac{25}{12\pi^2x^2} \left(\frac{sin(6\pi x/5)}{6\pi x/5} - cos(6\pi x/5) \right)}.
If \eqn{x=0/0} then the value \eqn{1} is given.
}
\details{
Note, this function is vectorized.
}
\examples{
quadratic_spectral_kernel(-20:20/15)
}
\references{
Horv\'{a}th, L., & Rice, G. (2024). Change Point Analysis for Time
Series (1st ed. 2024.). Springer Nature Switzerland.
\url{https://doi.org/10.1007/978-3-031-51609-2}
}
