% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funts.R
\name{quantile.funts}
\alias{quantile.funts}
\title{Quantile funts}
\usage{
\method{quantile}{funts}(x, probs = seq(0, 1, 0.25), ...)
}
\arguments{
\item{x}{A funts object}

\item{probs}{Numerics in [0,1] indicating the probabilities of interest}

\item{...}{Additional parameters to pass into quantile function}
}
\value{
Matrix with columns for each requested quantile
}
\description{
Obtain the pointwise quantile information of funts objects
}
\examples{
result <- quantile(funts(electricity))
result <- quantile(funts(electricity),probs = 0.95)
}
