% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{rainbow_plot}
\alias{rainbow_plot}
\title{Plot data as stack on 2d plot}
\usage{
rainbow_plot(object, CPs = NULL)
}
\arguments{
\item{object}{funts object}

\item{CPs}{CPs, if there are any. Default is NULL.}
}
\value{
Plot (ggplot2) with colored observations / regions
}
\description{
Plot data as stack on 2d plot
}
\examples{
rainbow_plot(funts(electricity))
rainbow_plot(funts(electricity), CPs=c(50,100,200))
}
