% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{rates}
\alias{rates}
\title{US Yield Curves}
\format{
\subsection{\code{rates}}{

A dfts object.
}
}
\usage{
rates
}
\description{
Yield curves in the US for 1 - 360 month maturity from 1990 to 2023
(removing days without information, i.e. weekends and holidays).
}
\keyword{datasets}
