% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_val_function.R
\name{required_val_function}
\alias{required_val_function}
\title{Potential required value function}
\usage{
required_val_function(data)
}
\arguments{
\item{data}{Numeric data.frame with rows for evaluated values and columns
indicating FD}
}
\value{
Numeric value indicating necessary value
}
\description{
Potential required value function
}
\examples{
required_val_function(electricity)
}
\references{
Rice, G, Zhang, C. (2021), \emph{Consistency of binary
segmentation for multiple change-point estimation with functional data}
(https://www.sciencedirect.com/science/article/pii/S0167715221001905)
}
