% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_change.R
\name{robust_change}
\alias{robust_change}
\title{Robust Change Point Detection}
\usage{
robust_change(X, m)
}
\arguments{
\item{m}{number of bootstrap repetitions}

\item{n}{length of observation}

\item{d}{dimension of observation}

\item{ht}{heavy tailed scenario (1) or normal-distribution (0)}

\item{S}{number of simulation runs}

\item{quantiles}{desired quantiles}
}
\value{
A list with the following elements
\itemize{
\item Wilcox_pvalue: p-value for the Wilcox test statistic
\item Cusum_pvalue: p-value for the cusum test statistic
\item Wilcox_bandwidth: Bandwidth for Wilcox test
\item Cusum_bandwidth: Bandwidth for cusum test
}
}
\description{
Robust Change Point Detection
}
\examples{
robust_change(funts(electricity),m=10)
}
\references{
Wegner, L., Wendler, M. Robust change-point detection for
functional time series based on U-statistics and dependent wild bootstrap.
Stat Papers (2024). https://doi.org/10.1007/s00362-024-01577-7
}
