% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facf_functions.R, R/tests_wn_ind_functions.R
\name{scalar_covariance_i_j}
\alias{scalar_covariance_i_j}
\title{Scalar Covariance for i,j}
\usage{
scalar_covariance_i_j(f_data, i, j, times)

scalar_covariance_i_j(f_data, i, j, times)
}
\arguments{
\item{f_data}{the functional data matrix with observed functions in the columns}

\item{i, j}{the indices i,j in 1:T that we are computing the covariance for}

\item{times}{A vector with 4 columns containing indices specifying which subset of f_data to consider}
}
\value{
A numeric value; the covariance of the functional data at a point for lag
windows defined by i,j (c^hat_i_j(t,s,u,v)).

A numeric value; the covariance of the functional data at a point for lag
windows defined by i,j.
}
\description{
Compute the approximate covariance at a point for lag windows defined by i,j

\code{scalar_covariance_i_j} Computes the approximate covariance at a point of the functional data
for lag windows defined by i,j; a scalarized version of covariance_i_j that takes point estimates.
}
\details{
\code{scalar_covariance_i_j} Computes the approximate covariance at a point of the functional data
for lag windows defined by i,j; a scalarized version of covariance_i_j that takes point estimates.
scalar_covariance_i_j returns the approximate covariance c^hat_i_j(t,s,u,v) evaluated at a
given t,s,u,v in U_J X U_J X U_J X U_J (for use in MCint method).
}
