% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar_detection.R
\name{scalarDetection}
\alias{scalarDetection}
\title{Scalar Characteristic Function Change Point Analysis}
\usage{
scalarDetection(Y, gam = 0.5, nSims = 200, alpha_val = 0.05)
}
\arguments{
\item{Y}{Scalar data}

\item{gam}{Numeric for weights}

\item{nSims}{Numeric for number of simulations}

\item{alpha_val}{Signficance of test}
}
\value{
XXXXX
}
\description{
This implements a scalar characteristic change method.
}
\examples{
scalarDetection(c(rnorm(50),rnorm(75,mean = 1)))
\donttest{
scalarDetection(as.numeric(colMeans(electricity)))
}
}
\references{
Hušková, M., & Meintanis, S.G. (2006). Change Point Analysis
based on Empirical Characteristic Functions. Metrika, 63, 145-168.
}
