% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{sd.funts}
\alias{sd.funts}
\title{SD for funts}
\usage{
\method{sd}{funts}(x, type = c("op", "pw"), ...)
}
\arguments{
\item{type}{Character to specify if an operator ('op') or pointwise ('pw')
calculation is desired}

\item{...}{Unused parameters}
}
\description{
SD for funts
}
\examples{
sd(funts(electricity),type='pw')
}
