% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulate_size}
\alias{simulate_size}
\title{Simulate Size tests}
\usage{
simulate_size(data_length = 100, dep = 0, nSims = 500, path = tempdir())
}
\arguments{
\item{data_length}{(Optional) Positive integer indicating the length of data.
Default is 100}

\item{dep}{(Optional) Numeric in \link{0,1} indicating the dependence strength between
successive functional observations Default is 0.}

\item{nSims}{(Optional) Positive integer indicating the number of simulations.
Default in 500.}

\item{path}{String indicates the place of .rds data file, saved in
simulate_data_size}
}
\value{
Data.frame with p-values for several change point methods.
}
\description{
This function uses simulated data to determine size for several methods
}
\examples{
simulate_size(100,0)
simulate_size(100,0.5)
}
