% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_segmentation.R
\name{single_binary_segmentation}
\alias{single_binary_segmentation}
\title{Single Binary Segmentation}
\usage{
single_binary_segmentation(
  data,
  changepoint_function,
  cutoff_function,
  trim_function,
  alpha = 0.05,
  include_value = FALSE,
  ...
)
}
\arguments{
\item{data}{Numeric data.frame with rows for evaluated values and columns
indicating functional observations}

\item{changepoint_function}{XXXXXXXXXXXXXXXXX.}

\item{cutoff_function}{XXXXXX}

\item{trim_function}{XXXXX}

\item{alpha}{Numeric value in \eqn{[0, 1]} indicating the significance for
cutoff_function.}

\item{include_value}{XXXXX}

\item{...}{Additional arguments passed into changepoint_function}
}
\value{
A numeric value indicating the cutoff location (if exists),
NA otherwise
}
\description{
This function implements traditional binary segmentation on functional data
for general functions. At most one change point is detected.
}
\examples{
single_binary_segmentation(
  data = electricity[, 1:60],
  changepoint_function = compute_Mn,
  cutoff_function = welch_approximation,
  trim_function = function(data) {
    max(10, floor(log(ncol(as.data.frame(data)))),
      na.rm = TRUE
    )
  }
)
}
