% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_measuring_vectors.R
\name{space_measuring_vectors}
\alias{space_measuring_vectors}
\title{Compute Spacing Measuring Vectors}
\usage{
space_measuring_vectors(X, M = 20, space = "BM")
}
\arguments{
\item{X}{A dfts object or data which can be automatically converted to that
format. See \code{\link[=dfts]{dfts()}}.}

\item{M}{Integer for the number of vectors to generate}

\item{space}{String for the space of interest. Options are Brownian motion
('BM'), OU process ('OU'), principal components ('PC'), and a vectors in iid
standard, random normals ('RN').}
}
\value{
Data.frame with columns of vectors describing the space. Columns are
independent vectors.
}
\description{
This function is used to compute vectors from spaces to measure the function
data objects.
}
\examples{
space_measuring_vectors(M=10, space="BM", X=electricity)
#space_measuring_vectors(M=10, space="OU", X=electricity)
space_measuring_vectors(M=10, space="PC", X=electricity)
}
