% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_eigen_change.R
\name{specific_eigen_change}
\alias{specific_eigen_change}
\title{Detecting Changes in the Eigenvalues of the Covariance Operator of the Functional Data}
\usage{
specific_eigen_change(
  X,
  component,
  h = 2,
  CPs = NULL,
  delta = 0.1,
  M = 1000,
  K = bartlett_kernel
)
}
\arguments{
\item{component}{The eigenvalue that the componentwise test is applied to.}

\item{h}{The window parameter for the estimation of the long run covariance matrix. The default
value is \code{h=2}.}

\item{delta}{Trimming parameter to estimate the covariance function using partial sum estimates.}

\item{M}{Number of monte carlo simulations used to get the critical values. The default value is \code{M=1000}}

\item{fdobj}{A functional data object of class '\code{fd}'}

\item{mean_change}{If \code{TRUE} then the data is centered considering the change in the mean function.}
}
\value{
\item{\code{pvalue}}{
Approximate p value for testing whether there is a significant change in the desired eigenvalue of the covariance operator
}
\item{\code{change}}{
Estimated change location
}
}
\description{
This function tests and detects changes in the specific eigenvalue of
the covariance operator.
}
\details{
This function dates and detects changes in the defined eigenvalue of the covariance function. The critical values are
approximated via \code{M} Monte Carlo simulations.
}
\examples{
specific_eigen_change(generate_brownian_motion(500),1)
specific_eigen_change(funts(electricity),1)
specific_eigen_change(funts(electricity),2)
}
\references{
Aue, A, G Rice, and O Sönmez. “Structural Break Analysis
for Spectrum and Trace of Covariance Operators.” Environmetrics
(London, Ont.) 31, no. 1 (2020). https://doi.org/10.1002/env.2617.
}
