% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_wn_ind_functions.R
\name{spectral_t_statistic}
\alias{spectral_t_statistic}
\title{Compute Spectral Density Test}
\usage{
spectral_t_statistic(f_data, kernel = "Bartlett", bandwidth = "adaptive")
}
\arguments{
\item{f_data}{The functional data matrix with observed functions in columns}

\item{kernel}{The kernel function to use. The currently supported kernels are 'Bartlett' and 'Parzen'.}

\item{bandwidth}{specifies the bandwidth to use. Currently admitted arguments are positive}
}
\value{
scalar value of the spectral density based test statistic.
}
\description{
Computes the spectral density operator based test statistic of the functional
data f_data.
}
