% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{stat_3D}
\alias{stat_3D}
\title{Draw 3D Geoms}
\usage{
stat_3D(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{geom}{The geom type to use \emph{e.g. "point", "path", or "line"}.}

\item{...}{Arguments passed on to layer. Often the aesthetics
like color = "red" or size = 3.}

\item{theta}{The azimuthal rotation (degrees).}

\item{phi}{The colatitude rotation (degrees).}
}
\description{
This function adds 3D geoms such as points and paths to a ggplot2 plot.
}
\examples{
dat <- funts(electricity)

data_lines <- cbind(data.frame('Time'=dat$intraobs), dat$data) \%>\%
 tidyr::pivot_longer(cols = 1+1:ncol(dat$data))

colors_plot <- RColorBrewer::brewer.pal(11, "Spectral")
colors_plot <- grDevices::colorRampPalette(colors_plot)(ncol(dat$data))
data_lines$color <- rep(colors_plot, nrow(dat$data) )
data_lines$name <- as.numeric(data_lines$name)

result <- ggplot2::ggplot(data_lines,
   ggplot2::aes(y=Time, x=name, z=value, color=color)) +
ggplot2::theme_void() +
stat_3D(theta=0, phi=15, geom='path') +
ggplot2::scale_color_manual(
   breaks = data_lines$color,
   values = data_lines$color
) +
ggplot2::guides(color='none')
}
\references{
Acker D (2024). gg3D: 3D perspective plots for ggplot2. R
package version 0.0.0.9000.
}
