% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationarity_test.R
\name{stationarity_test}
\alias{stationarity_test}
\title{Functional Stationarity Test}
\usage{
stationarity_test(
  X,
  statistic = "Tn",
  critical = c("simulation", "permutation"),
  perm_method = "separate",
  M = 1000,
  blocksize = 3,
  TVE = 1,
  replace = TRUE
)
}
\arguments{
\item{X}{A dfts object or data which can be automatically converted to that
format. See \code{\link[=dfts]{dfts()}}.}

\item{statistic}{String for test statistic. Options are \code{Tn} and
\code{Mn}. Default is \code{Tn}.}

\item{critical}{String for method of determining the critical values. Options are
\code{simulation} and \code{permutation}. Default is \code{simulation}.}

\item{perm_method}{String for method of permutation. Options are \code{separate}
and \code{sliding}. Default is \code{separate}.}

\item{M}{Numeric for number of simulation to use in determining the null
distribution. Default is 1000.}

\item{blocksize}{Numeric for blocksize in permutation test. Default is 1.}

\item{TVE}{Numeric for total variance explained when using PCA for
eigenvalues. Default is 1.}

\item{replace}{Boolean if replacement should be used for permutation test.
Default is TRUE.}
}
\value{
List with (1) pvalue, (2) test statistic, and (3) test statistics
from simulating the null distribuion.
}
\description{
Stationarity test for functional time series with different methods on
determining the critical values of the test statistic. The Monte Carlo
method was constructed in Horvath et al. (2014), while the bootstrap-based
methods have not been validated in the literature (although such an option
is provided, please use them at your own risk).
}
\examples{
res <- stationarity_test(
  generate_brownian_motion(100,v=seq(0,1,length.out=20)),
  critical='permutation', statistic='Mn')
res2 <- stationarity_test(electricity)
}
\references{
Horvath, L., Kokoszka, P., & Rice, G. (2014). Testing
stationarity of functional time series. Journal of Econometrics,
179(1), 66-82.
}
