% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.dfts}
\alias{summary.dfts}
\title{Summary for dfts Object}
\usage{
\method{summary}{dfts}(object, changes = NULL, lag.max = 20, d.max = 2, demean = FALSE, ...)
}
\arguments{
\item{object}{A dfts object or data which can be automatically converted to that
format. See \code{\link[=dfts]{dfts()}}.}

\item{changes}{changes, if there are any. Default is NULL.}

\item{lag.max}{Max lags to consider for ACF. Default is 20.}

\item{d.max}{Max number of dimensions for QQ-plot.}

\item{demean}{Boolean if data should be demeaned based on changes and create
plots based on these residuals.}

\item{...}{Unused.}
}
\value{
Plot (ggplot) summarizing the data
}
\description{
Summary for dfts Object
}
\examples{
res <- summary(dfts(electricity$data[,1:20]), lag.max=2)
# res1 <- summary(dfts(electricity$data[,1:20]), changes=c(3), lag.max=2)
# summary(electricity, changes=c(50,200,300), lag.max=5, demean = TRUE)
}
