% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facf_functions.R, R/tests_wn_ind_functions.R
\name{t_statistic_Q}
\alias{t_statistic_Q}
\title{FACF - Test Statistic}
\usage{
t_statistic_Q(f_data, lag)

t_statistic_Q(f_data, lag)
}
\arguments{
\item{f_data}{the functional data matrix with observed functions in columns}

\item{lag}{the fixed time lag used in the computation of the statistic}
}
\value{
scalar value of the statistic Q_{T,h} to test the hypothesis H_{0,h} : y_h(t,s) = 0.

scalar value of the statistic Q_{T,h} to test the hypothesis H_{0,h} : y_h(t,s) = 0.
}
\description{
Computes the test statistic Q_{T,h} = T*||y^hat_h||^2 for fixed h and for T
inferred from the functional data f_data that is passed.

Computes the test statistic Q_{T,h} = T*||y^hat_h||^2 for fixed h and for T
inferred from the functional data f_data that is passed.
}
