% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_wn_ind_functions.R
\name{t_statistic_V}
\alias{t_statistic_V}
\title{Test Statistic - t_statistic_V}
\usage{
t_statistic_V(f_data, K)
}
\arguments{
\item{f_data}{the functional data with functions in columns}

\item{K}{the max value in the range of time lags (1:K) used}
}
\value{
scalar value of the statistic V_{T,K} to test the hypothesis
H'_{0,K} : for all h in 1:K y_h(t,s) = 0.
}
\description{
Computes the statistic \code{V_{T,K} = T*sum_h(||y^hat_h||^2)} or h in 1:K and for T
inferred from the functional data f_data that is passed to the function.
}
