% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{temperature}
\alias{temperature}
\title{Australian Temperature Data}
\format{
\subsection{\code{temperature}}{

A list with 8 elements, one for each city. Each element is a data.frame with
365 rows (one per day), and a variable number of columns related to years of
collected data.
}
}
\source{
<www.bom.gov.au>
}
\usage{
temperature
}
\description{
A list of daily minimum temperature data from reporting stations in
Australia. Any missing data was previously, linearly, interpolated and leap
days were removed for consistency. Moreover, the data is classified by city
as only one station was observed in each city.
}
\keyword{datasets}
