% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_change.R
\name{tmp}
\alias{tmp}
\title{Need to Do:
\itemize{
\item Find and apply: sde and tensorA.
\item Convert = to <- (styler may handle it..)
\item Update documentation
\item Remove any functions not needed
}}
\usage{
tmp()
}
\description{
Need to Do:
\itemize{
\item Find and apply: sde and tensorA.
\item Convert = to <- (styler may handle it..)
\item Update documentation
\item Remove any functions not needed
}
}
