% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_change.R
\name{TNstat}
\alias{TNstat}
\title{Compute Covariance Tn Statistic}
\usage{
TNstat(xf)
}
\arguments{
\item{xf}{Data.frame of numerics. Finite realization of functional time
series data, where curves are stored in columns.}
}
\value{
XXXXX
}
\description{
This function computes the Tn statistic in the paper, introducde after
theorem 2.1
}
\examples{
# XXXXX
}
