% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_function.R
\name{trim_function}
\alias{trim_function}
\title{Basic trim function}
\usage{
trim_function(data, minVal = 10, ...)
}
\arguments{
\item{data}{Numeric data.frame with rows for evaluated values and columns
indicating FD}

\item{minVal}{Integer indicating the minimal trim amount}

\item{...}{Unused, but exist to allow for use in other functions}
}
\value{
Integer value indicating the amount to trim
}
\description{
Basic trim function to select how much data to consider.
}
\examples{
trim_function(electricity)
}
\references{
Rice, G, Zhang, C. (2021), \emph{Consistency of binary
segmentation for multiple change-point estimation with functional data}
(https://www.sciencedirect.com/science/article/pii/S0167715221001905)
}
