% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_wn_ind_functions.R
\name{true_eta_approx_i_j}
\alias{true_eta_approx_i_j}
\title{Approximate eta_i_j}
\usage{
true_eta_approx_i_j(f_data, i, j)
}
\arguments{
\item{f_data}{the functional data matrix with functions in columns}

\item{i, j}{the indices i,j in 1:T that we are computing eta^hat_i_j for}
}
\value{
scalar value of eta^hat_i_j computed using a simple Riemann sum.
}
\description{
a non-stochaistic approximation of eta_i_j (see (15)) using a Riemann sum.
}
