% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{truncated_kernel}
\alias{truncated_kernel}
\title{Truncated Kernel}
\usage{
truncated_kernel(x)
}
\arguments{
\item{x}{A numeric value at which to evaluate kernel.
It often indicates current lag divided by window.}
}
\value{
A numeric value indicating value of Kernel at that point
}
\description{
Kernel where \eqn{1, |x|\leq 1} and \eqn{0} otherwise. If \eqn{x=0/0} then
the value \eqn{1} is given.
}
\details{
Note, this function is vectorized.
}
\examples{
truncated_kernel(-20:20/15)
}
\references{
Horváth, L., & Rice, G. (2024). Change Point Analysis for Time
Series (1st ed. 2024.). Springer Nature Switzerland.
\url{https://doi.org/10.1007/978-3-031-51609-2}
}
