% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funts.R
\name{validate.funts}
\alias{validate.funts}
\title{Validate funts}
\usage{
validate.funts(x)
}
\arguments{
\item{x}{funts object to verify for any violations}
}
\value{
Silently returns the data if there are no violations
}
\description{
Validate funts
}
\examples{
validate.funts(funts(electricity))
}
