% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{var.funts}
\alias{var.funts}
\title{Variance for funts}
\usage{
\method{var}{funts}(x, type = c("op", "pw"), ...)
}
\arguments{
\item{type}{Character to specify if an operator ('op') or pointwise ('pw')
calculation is desired}

\item{...}{Unused parameters}
}
\description{
Variance for funts
}
\examples{
var(funts(electricity),type='pw')
}
