% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facf_functions.R, R/tests_wn_ind_functions.R
\name{variance_hat_Q_h}
\alias{variance_hat_Q_h}
\title{FACF - Welch Approximation Var Hat}
\usage{
variance_hat_Q_h(f_data, lag, M = NULL, low_disc = FALSE)

variance_hat_Q_h(f_data, lag, M = NULL, low_disc = FALSE)
}
\arguments{
\item{f_data}{the functional data matrix with functions in columns}

\item{lag}{specifies the lag use in the test statistic Q_h (lag = h in paper)}

\item{M}{optional argument specifying the sampling size in the related Monte Carlo method}

\item{low_disc}{boolean value specifiying whether or not to use low-discrepancy sampling
for the Monte-Carlo method (only Sobol Sampling is currently supported)}
}
\value{
Scalar approximation of the variance of the test statistic Q_h

scalar approximation of the variance of the test statistic Q_h
}
\description{
Computes the approximation of the variance defined in (15) which is used in
the Welch- Satterthwaite approximation as variance of the chi-squared random variable
approximating Q_h.

Computes the approximation of the variance defined in (15) which is used in
the Welch-Satterthwaite approximation as variance of the chi-squared random variable
approximating Q_h.
}
