% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facf_functions1.R, R/tests_wn_ind_functions.R
\name{variance_hat_Q_h_iid}
\alias{variance_hat_Q_h_iid}
\title{Test Statistic - Variance Q_h_iid}
\usage{
variance_hat_Q_h_iid(f_data)

variance_hat_Q_h_iid(f_data)
}
\arguments{
\item{f_data}{the functional data matrix with functions in columns}
}
\value{
scalar approximation of the variance of the test statistic Q_h

scalar approximation of the variance of the test statistic Q_h
}
\description{
Computes the approximation of the variance defined in (15) which is used
in the Welch- Satterthwaite approximation under the assumption that the functional data
follows a strong white noise.

Computes the approximation of the variance defined in (15) which is used
in the Welch- Satterthwaite approximation under the assumption that the functional data
follows a strong white noise.
}
