% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_wn_ind_functions.R
\name{variance_hat_V_K}
\alias{variance_hat_V_K}
\title{Variance Approximation - variance_hat_V_K}
\usage{
variance_hat_V_K(f_data, K, M = NULL, low_disc = FALSE)
}
\arguments{
\item{f_data}{the functional data matrix with functions in columns}

\item{K}{specifies the range of lags 1:K for the test statistic V_K}

\item{M}{optional argument specifying the sampling size in the related Monte Carlo method}

\item{low_disc}{boolean value specifiying whether or not to use low-discrepancy sampling
for the Monte-Carlo method (only Sobol Sampling is currently supported)}
}
\value{
scalar approximation of the variance of the test statistic V_K.
}
\description{
Computes the approximation of the variance defined in (15) which is used in
the Welch- Satterthwaite approximation as the variance of the chi-squared random variable
approximating V_K.
}
