% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_wn_ind_functions.R
\name{variance_hat_V_K_iid}
\alias{variance_hat_V_K_iid}
\title{Variance Approximation - iid variance_hat_V_K}
\usage{
variance_hat_V_K_iid(f_data, K)
}
\arguments{
\item{f_data}{the functional data matrix with functions in columns}

\item{K}{specifies the range of lags 1:K for the test statistic V_K}
}
\value{
scalar approximation of the variance of the test statistic V_K
}
\description{
Computes the approximation of the variance defined in (15) which is used
in the Welch-Satterthwaite approximation under the assumption that the functional data
follows a strong white noise.
}
