% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_change.R
\name{weight_criticalvalueMC}
\alias{weight_criticalvalueMC}
\title{Compute critical values for weight_TNstat ( \eqn{T_N(\kappa)} )}
\usage{
weight_criticalvalueMC(xf, len, kappa)
}
\arguments{
\item{xf}{XXXX}

\item{len}{XXXX}

\item{kappa}{XXXX}
}
\value{
XXXX
}
\description{
Compute critical values for weight_TNstat ( \eqn{T_N(\kappa)} )
}
\examples{
# XXXXX
}
