% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_change.R
\name{weight_TNstat}
\alias{weight_TNstat}
\title{Compute the Weight Tn Statistic}
\usage{
weight_TNstat(xf, kappa)
}
\arguments{
\item{xf}{XXXXX}

\item{kappa}{XXXXX}
}
\value{
XXXXX
}
\description{
The function computes the weighted Tn statistic, introduced after Theorem 2.3
}
\examples{
# XXXXX
}
