% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welch_approximation.R
\name{welch_approximation}
\alias{welch_approximation}
\title{Welch Approximation to Tn}
\usage{
welch_approximation(
  X,
  alpha = 0.05,
  TVal = length(X[1, ]),
  W = NULL,
  W1 = NULL,
  M = 100,
  h = TVal^(1/3),
  K = bartlett_kernel,
  ...
)
}
\arguments{
\item{X}{Numeric data.frame with evaled points on rows and fd objects in columns}

\item{alpha}{(Optional) Numeric value in (0, 1) for significance. Default is 0.05}

\item{TVal}{(Optional) Numeric value indicate the number of FDs. Default is the
number of columns in X.}

\item{W}{(Optional) Data.frame for the functions to integrate against
used for both muHat and sigmaHat2. Default of NULL will use Brownian
motion.}

\item{W1}{(Optional)  data.frame for the functions to integrate against
used for sigmaHat2. Default of NULL will use Brownian motion.}

\item{M}{(Optional) Numeric value for number of functions to integrate against
(ncol(W)). Default is 100.}

\item{h}{(Optional) Numeric value indicating the amount of blocking. Default
is Tval^(1/3)}

\item{K}{(Optional) Function used for CHat, indicating Kernel. Default is the
Bartlett Kernel.}

\item{...}{Unused parameters, but added for use in other functions.}
}
\value{
Numeric value indicating cutoff from Welsh approximation
}
\description{
This function approximates the Tn statistic using the Welch approximation.
}
\examples{
# Note, TVal generally bigger than 1.
welch_approximation(electricity, TVal = 1, h=0)
}
