% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welsh_approximation.R
\name{welsh_approximation}
\alias{welsh_approximation}
\title{welsh Approximation to Tn}
\usage{
welsh_approximation(
  X,
  alpha = 0.05,
  TVal = length(X[1, ]),
  W = NULL,
  W1 = NULL,
  M = 100,
  h = TVal^(1/3),
  K = bartlett_kernel,
  ...
)
}
\arguments{
\item{X}{Numeric data.frame with evaled points on rows and fd objects in columns}

\item{alpha}{Numeric value in (0,1) for significance}

\item{TVal}{Numeric value indicate the number of FDs}

\item{W}{Optional data.frame for the functions to integrate against
used for both muHat and sigmaHat2}

\item{W1}{Optional data.frame for the functions to integrate against
used for sigmaHat2}

\item{M}{Numeric value for number of functions to integrate against (ncol(W))}

\item{h}{Numeric value indicating the amount of blocking}

\item{K}{Function used for CHat}
}
\value{
Numeric value indicating cutoff
}
\description{
welsh Approximation to Tn
}
\examples{
# Setup Data
data_KL <- generate_data_fd(ns = c(100,100),
    eigsList = list(c(3,2,1,0.5),
                    c(3,2,1,0.5)),
    basesList = list(fda::create.bspline.basis(nbasis=4, norder=4),
                     fda::create.bspline.basis(nbasis=4, norder=4)),
    meansList = c(-1,1),
    distsArray = c('Normal','Normal'),
    evals = seq(0,1,0.05),
    kappasArray = c(0,0))
welsh_approximation(data_KL)
}
