% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_change.R
\name{ZNstat}
\alias{ZNstat}
\title{Compute Zn Statistic for Funtional Covariance Changes}
\usage{
ZNstat(xdm, u)
}
\arguments{
\item{xdm}{Data.frame of numerics. It is the finite realization of DEMEANed
functional time series data, where curves are stored in columns.}

\item{u}{Numeric in \eqn{(0, 1)}. That is, a fraction index over the interval
\eqn{[0, 1]}.}
}
\value{
XXXXX
}
\description{
This (internal) function implements Theorem 2.1 / 2.2.
}
\examples{
# XXXXX
}
