% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_change.R
\name{ZNstat_cp}
\alias{ZNstat_cp}
\title{Title}
\usage{
ZNstat_cp(xdm, u)
}
\arguments{
\item{xdm}{Data.frame of numerics. It is the finite realization of DEMEANed
functional time series data, where curves are stored in columns.}

\item{u}{Numeric in \eqn{(0, 1)}. That is, a fraction index over the interval
\eqn{[0, 1]}.}
}
\value{
XXXXX
}
\description{
Title
}
\examples{
# XXXXX
}
