% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_wn_ind_functions.R
\name{adaptive_bandwidth}
\alias{adaptive_bandwidth}
\title{Adaptive_bandwidth}
\usage{
adaptive_bandwidth(f_data, kernel)
}
\arguments{
\item{f_data}{the functional data matrix with observed functions in columns}

\item{kernel}{the kernel function to use. The currently supported kernels are 'Bartlett' and 'Parzen'.
The default kernel is 'Bartlett'.}
}
\value{
a scalar value of the "optimal" data-adapted bandwidth.
}
\description{
Computes the "optimal" bandwidth using a bandwidth selection method based on the
spectral density operator which adapts to the functional data.
}
