% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{adaptive_bandwidth}
\alias{adaptive_bandwidth}
\title{Adaptive_bandwidth}
\usage{
adaptive_bandwidth(X, kernel = "bartlett")
}
\arguments{
\item{X}{A dfts object or data which can be automatically converted to that
format. See \code{\link[=dfts]{dfts()}}.}

\item{kernel}{Kernel function selection to use, 'bartlett' or 'parzen'.}
}
\value{
a scalar value of the "optimal" data-adapted bandwidth.
}
\description{
Computes the "optimal" bandwidth using a bandwidth selection method based on the
spectral density operator which adapts to the functional data.
}
