% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funts.R
\name{average}
\alias{average}
\alias{mean.funts}
\alias{median.funts}
\title{Average Functions for funts Objects}
\usage{
\method{mean}{funts}(x, na.rm = TRUE, ...)

\method{median}{funts}(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{funts object}

\item{na.rm}{Boolean if NA values should be removed. Defaults to TRUE.}

\item{...}{Additional parameters to stats function}
}
\value{
Numeric vector
}
\description{
Compute the average values for funts objects
}
\examples{
results <- mean(funts(electricity))
results <- median(funts(electricity))
}
