% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welsh_approximation.R
\name{bartlett_kernel}
\alias{bartlett_kernel}
\title{Bartlett Kernel}
\usage{
bartlett_kernel(l, h)
}
\arguments{
\item{l}{A numeric value indicating current lag}

\item{h}{A numeric value indicating smallest lag with no influence}
}
\value{
A numeric value indicating value of Kernel at that point
}
\description{
One of the most obvious options for the K function in the Welsh-approximation
is the Barlett Kernel.
}
\examples{
bartlett_kernel(1,2)

# Use in Welsh Approximation
data_KL <- generate_data_fd(ns = c(100,100),
    eigsList = list(c(3,2,1,0.5),
                    c(3,2,1,0.5)),
    basesList = list(fda::create.bspline.basis(nbasis=4, norder=4),
                     fda::create.bspline.basis(nbasis=4, norder=4)),
    meansList = c(-1,1),
    distsArray = c('Normal','Normal'),
    evals = seq(0,1,0.05),
    kappasArray = c(0,0))
welsh_approximation(data_KL,K=bartlett_kernel)
}
