% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{bartlett_kernel}
\alias{bartlett_kernel}
\title{Bartlett Kernel}
\usage{
bartlett_kernel(x)
}
\arguments{
\item{x}{A numeric value at which to evaluate kernel.
It often indicates current lag divided by window.}
}
\value{
A numeric value indicating value of Kernel at that point
}
\description{
Kernel where \eqn{max(0,1-|m/h|), h\neq 0} and \eqn{1, h=0}.
}
\details{
Note, this function is vectorized.
}
\examples{
bartlett_kernel(-20:20/15)
}
