% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{center}
\alias{center}
\alias{center.default}
\alias{center.data.frame}
\alias{center.matrix}
\alias{center.funts}
\title{Generic Centering of Data}
\usage{
center(x, ...)

\method{center}{default}(x, ...)

\method{center}{data.frame}(x, ...)

\method{center}{matrix}(x, ...)

\method{center}{funts}(x, CPs = NULL, ...)
}
\arguments{
\item{...}{Additional parameters to function based on data}

\item{object}{Object for computation of centering}
}
\value{
Centered data of the same format as the given data
}
\description{
Generic Centering of Data

Center funts data
}
\examples{
center(1:10)
}
