% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{center}
\alias{center}
\alias{center.default}
\alias{center.data.frame}
\alias{center.matrix}
\alias{center.funts}
\title{Generic Centering of Data}
\usage{
center(object, CPs = NULL, type = "mean", ...)

\method{center}{default}(object, CPs = NULL, type = "mean", ...)

\method{center}{data.frame}(object, CPs = NULL, type = "mean", ...)

\method{center}{matrix}(object, CPs = NULL, type = "mean", ...)

\method{center}{funts}(object, CPs = NULL, type = "mean", ...)
}
\arguments{
\item{object}{Object for computation of centering}

\item{CPs}{Change points for centering individual sections.}

\item{type}{String of \code{mean} or \code{median} for method of centering.}

\item{...}{Unused}
}
\value{
Centered data of the same format as the given data
}
\description{
Generic Centering of Data
}
\examples{
center(1:10)
}
\seealso{
\code{\link[=center.default]{center.default()}}, \code{\link[=center.data.frame]{center.data.frame()}},
\code{\link[=center.matrix]{center.matrix()}}, \code{\link[=center.funts]{center.funts()}}
}
