% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{center}
\alias{center}
\alias{center.default}
\alias{center.data.frame}
\alias{center.matrix}
\alias{center.dfts}
\title{Generic Centering of Data}
\usage{
center(object, changes = NULL, type = "mean", ...)

\method{center}{default}(object, changes = NULL, type = "mean", ...)

\method{center}{data.frame}(object, changes = NULL, type = "mean", ...)

\method{center}{matrix}(object, changes = NULL, type = "mean", ...)

\method{center}{dfts}(object, changes = NULL, type = "mean", ...)
}
\arguments{
\item{object}{Object for computation of centering.}

\item{changes}{Change points for centering individual sections.}

\item{type}{String of \code{mean} or \code{median} for method of centering.}

\item{...}{Parameters that may be fed into other versions of centering.}
}
\value{
Centered data of the same format as the given data.
}
\description{
Center data by removing the mean or median. Defining changes allow for
regional centering.
}
\examples{
center(1:10)
}
\seealso{
\code{\link[=center.default]{center.default()}}, \code{\link[=center.data.frame]{center.data.frame()}},
\code{\link[=center.matrix]{center.matrix()}}, \code{\link[=center.dfts]{center.dfts()}}
}
