% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_segmentation.R
\name{complete_binary_segmentation}
\alias{complete_binary_segmentation}
\title{Complete Binary Segmentation}
\usage{
complete_binary_segmentation(
  data,
  changepoint_function = compute_Mn(),
  cutoff_function = welch_approximation(),
  final_verify = TRUE,
  silent = FALSE,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{data}{Numeric data.frame with rows for evaluated values and columns
indicating functional observations}

\item{changepoint_function}{XXXXXXXXXXXXXXXXX.}

\item{cutoff_function}{XXXXXX}

\item{final_verify}{(Optional) Boolean value XXXXXXXXXXXXX

Indicates if a final pass looking at sequences with only one change point
should be conducted to verify results. Note, this may modify existing
locations of change points, potentially to less accurate locations.}

\item{silent}{(Optional) Boolean value

Indicates if useful output should be silenced. Default FALSE shows output.}

\item{alpha}{Numeric value in \eqn{[0, 1]} indicating the significance for
cutoff_function.}

\item{...}{Additional arguments passed into changepoint_function}
}
\value{
A list of numeric values indicating change points  (if exists),
NA otherwise
}
\description{
This function implements traditional binary segmentation on functional data
for general functions. Change points are recursively found until no
more change points are detected.
}
\examples{
complete_binary_segmentation(data = electricity[,1:80],
  changepoint_function = compute_Mn,
  cutoff_function = welch_approximation,
  trim_function = function(data){
    max(50, floor(log(ncol(as.data.frame(data)))),
        na.rm=TRUE)},
  final_verify=FALSE)
}
