% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistics.R
\name{compute_Mn}
\alias{compute_Mn}
\title{Compute Mn Test Statistic}
\usage{
compute_Mn(X, M = 10000, W = NULL, space = "BM", ...)
}
\arguments{
\item{X}{Numeric data.frame with rows for evaluated values and columns
indicating FD.}

\item{M}{(Optional) Numeric indicating the number of vectors used to span W.
Defaults to 10,000.}

\item{W}{(Optional) Data.frame of numerics with rows for evaluated values
and columns indicating function that contains the vectors to span the
space. Note the number of columns should match that of X. Default is NULL,
which generates the vectors in function. This overrides space and M if
not NULL.}

\item{space}{(Optional) String indicating the space to integrate against.
Default is "BM", or Brownian motion. See \code{computeSpaceMeasuringVectors()}
for more information.}

\item{...}{Parameters are not passed anywhere. Just added for use in other
functions.}
}
\value{
A list with three elements:
\itemize{
\item \strong{value}: Numeric for maximum Mn value in data.
\item \strong{location}: Numeric for location of maximum in data.
\item \strong{allValues}: Vector of numerics for Mn values at each point.
}
}
\description{
Function to calculate the change point test statistic for functional data
\deqn{Mn = sup_{x\in (0, 1)} \int |Z_n(v,x)|^2 dQ(v)}
}
\examples{
compute_Mn(electricity)
}
