% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_detection.R
\name{compute_Mn}
\alias{compute_Mn}
\title{Compute Characteristic Maximized Test Statistic}
\usage{
compute_Mn(
  X,
  W = computeSpaceMeasuringVectors(M = 20, X = X, space = "BM"),
  J = min(50, ncol(X))
)
}
\arguments{
\item{X}{Data.frame of the observations}

\item{J}{Numeric for the reslolution to check}
}
\value{
List with the following
\itemize{
\item value: numeric maximized test statistic
\item location: placement of maximized value
\item allValues: test statistic value at each point
}
}
\description{
Computes the maximized test statistic for the characteristic functional-based
change point detection.
}
\examples{
compute_Mn(electricity[,1:100])
}
