% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistics.R
\name{compute_Tn}
\alias{compute_Tn}
\title{Compute Tn Test Statistic}
\usage{
compute_Tn(X, k = NULL, M = 100, W = NULL, ...)
}
\arguments{
\item{X}{Numeric data.frame with rows for evaluated values and columns
indicating FD}

\item{k}{(Optional) Defaults to NULL, but can be an integer in \link{1,ncol(X)}

Indicates the candidate change point to investigate. If k=NULL then get the
test statistic for entire sample}

\item{M}{(Optional) Defaults to 100. The number of functions in W}

\item{W}{(Optional) Optional numeric matrix with rows for evaluated values
and columns indicating function

This is the respective functions used to investigate the test statistic. The
NULL value means Brownian motion is used.}
}
\value{
A numeric value for the test statistic for entire sample or candidate
change point.
}
\description{
Function to calculate the change point test statistic for functional data
\deqn{Tn = \int_0^1 \int |Z_n(v,x)|^2 dQ(v) dx}
}
\examples{
# Setup Data
data_KL <- generate_data_fd(ns = c(200),
    eigsList = list(c(3,2,1,0.5)),
    basesList = list(fda::create.bspline.basis(nbasis=4, norder=4)),
    meansList = c(0),
    distsArray = c('Normal'),
    evals = seq(0,1,0.05),
    kappasArray = c(0))
compute_Tn(data_KL)

# Setup Data
data_KL <- generate_data_fd(ns = c(100,100),
    eigsList = list(c(3,2,1,0.5),
                    c(3,2,1,0.5)),
    basesList = list(fda::create.bspline.basis(nbasis=4, norder=4),
                     fda::create.bspline.basis(nbasis=4, norder=4)),
    meansList = c(-0.5,0.5),
    distsArray = c('Normal','Normal'),
    evals = seq(0,1,0.05),
    kappasArray = c(0,0))
compute_Tn(data_KL)
}
