% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistics.R
\name{compute_Tn}
\alias{compute_Tn}
\title{Compute Tn Test Statistic}
\usage{
compute_Tn(X, M = 1e+05, W = NULL, space = "BM", ...)
}
\arguments{
\item{X}{Numeric data.frame with rows for evaluated values and columns
indicating FD.}

\item{M}{(Optional) Numeric indicating the number of vectors used to span W.
Defaults to 10,000.}

\item{W}{(Optional) Data.frame of numerics with rows for evaluated values
and columns indicating function that contains the vectors to span the
space. Note the number of columns should match that of X. Default is NULL,
which generates the vectors in function. This overrides space and M if
not NULL.}

\item{space}{(Optional) String indicating the space to integrate against.
Default is "BM", or Brownian motion. See \code{computeSpaceMeasuringVectors()}
for more information.}

\item{...}{Parameters are not passed anywhere. Just added for use in other
functions.}
}
\value{
Numeric value for the test statistic of entire sample or candidate
change point.
}
\description{
Function to calculate the change point test statistic for functional data
\deqn{Tn = \int_0^1 \int |Z_n(v,x)|^2 dQ(v) dx}
}
\examples{
compute_Tn(electricity, M = 1000)
}
