% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_kernel_change.R
\name{covariance_kernel_change}
\alias{covariance_kernel_change}
\title{Covariance Kernel Change}
\usage{
covariance_kernel_change(X, kappa = 1/4, len = 30)
}
\arguments{
\item{X}{Numeric data.frame of functional data observations--rows for
evaluated values and columns indicating FD}

\item{kappa}{(Optional) Numeric used for weighting and such. Default is 1/4.}

\item{len}{(Optional) Numeric for window/repetitions for covariance change.
Default is 30.}
}
\value{
Location of change. NA for no change and numeric if there is a change.
}
\description{
This method implements a method for detection of changes in the covariance
kernel of functional data based on the norms of a generally weighted
process of partial sample estimates.
}
\details{
Upcoming: Size estimates may be included (already coded).
}
\examples{
result <- covariance_kernel_change(electricity[,1:18], len=20)
}
\references{
Horváth, L., Rice, G., & Zhao, Y. (2022). Change point analysis
of covariance functions: A weighted cumulative sum approach. Journal of
Multivariate Analysis, 189, 104877-.
\url{https://doi.org/10.1016/j.jmva.2021.104877}

https://github.com/yzhao7322/CovFun_Change
}
