% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{daniell_kernel}
\alias{daniell_kernel}
\title{Daniell Kernel}
\usage{
daniell_kernel(x)
}
\arguments{
\item{x}{A numeric value at which to evaluate kernel.
It often indicates current lag divided by window.}
}
\value{
A numeric value indicating value of Kernel at that point
}
\description{
Kernel where \eqn{sin(pi * x) / (pi * x), x!=0} and \eqn{1,X\x=0}.
}
\details{
Note, this function is vectorized.
}
\examples{
daniell_kernel(-20:20/15)
}
