% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_bootstrap.R
\name{detect_changepoint_bootstrap}
\alias{detect_changepoint_bootstrap}
\title{Detect Change Point - Bootstrap}
\usage{
detect_changepoint_bootstrap(
  X,
  statistic = c("Tn", "Mn"),
  M = 20,
  J = 50,
  space = "BM",
  blockSize = 1,
  iters = 1000,
  replace = FALSE,
  alpha = 0.05,
  silent = FALSE
)
}
\arguments{
\item{X}{Data.frame of the data}

\item{statistic}{Test statistic (Tn or Mn)}

\item{silent}{}
}
\description{
Detect the changes in data by using bootstrap.
}
