% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_bootstrap.R
\name{detect_changepoint_bootstrap}
\alias{detect_changepoint_bootstrap}
\title{Detect Change Point - Bootstrap}
\usage{
detect_changepoint_bootstrap(
  X,
  statistic = c("Tn", "Mn"),
  M = 20,
  J = 50,
  space = "BM",
  blockSize = 1,
  iters = 1000,
  replace = FALSE,
  alpha = 0.05,
  silent = FALSE
)
}
\arguments{
\item{X}{Data.frame of the data}

\item{statistic}{Test statistic (Tn or Mn)}

\item{M}{Number of vectors to explore space}

\item{J}{Resolution of curves for discretazation}

\item{space}{Numeric to indicate space to measure data in}

\item{blockSize}{Numeric for blocklength to manage dependece}

\item{iters}{Numeric for number of iteration for bootstrapped statistic}

\item{replace}{Boolean to convert from bootstrap to permutation}

\item{alpha}{Numeric for significance level}

\item{silent}{Boolean if output should be given while running}
}
\value{
List with the following parameters:
'value': Value of the test statistic for the original data
'cutoff': Cutoff from the bootstrap sample
'pval': p-value for the placement of the true in the bootstrapped sample
'BSSamples': Bootstrap samples
}
\description{
Detect the changes in data by using bootstrap.
}
\examples{
detect_changepoint_bootstrap(electricity,statistic = 'Tn')
}
