% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_detection.R
\name{detect_changepoint_final_Tn}
\alias{detect_changepoint_final_Tn}
\title{Title}
\usage{
detect_changepoint_final_Tn(
  X,
  M = 20,
  J = 50,
  nSims = 1000,
  h = 3,
  K = bartlett_kernel,
  space = "BM",
  silent = FALSE
)
}
\arguments{
\item{silent}{}
}
\description{
Title
}
