% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welch_approximation.R
\name{detect_changepoint_Welch}
\alias{detect_changepoint_Welch}
\title{Title}
\usage{
detect_changepoint_Welch(
  X,
  M = 20,
  J = 50,
  h = ncol(X)^(1/3),
  K = bartlett_kernel,
  space = "BM",
  alpha = 0.05
)
}
\arguments{
\item{alpha}{}
}
\description{
Title
}
