% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfts.R, R/generic_functions.R
\name{dfts_group}
\alias{dfts_group}
\alias{Math.dfts}
\alias{Ops.dfts}
\alias{cumsum.dfts}
\title{Group Generic Functions}
\usage{
\method{Math}{dfts}(x, ...)

\method{Ops}{dfts}(e1, e2)

\method{cumsum}{dfts}(x)
}
\arguments{
\item{x, e1, e2}{A dfts object. See \code{\link[=dfts]{dfts()}}.}

\item{...}{Further arguments passed to the methods.}
}
\value{
A dfts object with the applied operation

dfts object with data as cumsum
}
\description{
Group generic methods defined for things like Math, Ops, and so forth.
}
\examples{
result <- sqrt( electricity )
result <- electricity + electricity
result1 <- electricity * electricity
cumsum(electricity)
}
