% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funts.R
\name{diff.funts}
\alias{diff.funts}
\title{Difference funts}
\usage{
\method{diff}{funts}(x, lag = 1L, differences = 1L, ...)
}
\arguments{
\item{x}{funts object}

\item{lag}{An integer indicating which lag to use.}

\item{differences}{An integer indicating the order of the difference.}

\item{...}{Further arguments to be passed to or from methods.}
}
\value{
funts object with differenced values
}
\description{
Difference funts
}
\examples{
# diff(funts(electricity), lag=1)
# diff(funts(electricity), differences=2)
}
