% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funts.R
\name{dim.funts}
\alias{dim.funts}
\title{Dimension of funts}
\usage{
\method{dim}{funts}(x, ...)
}
\arguments{
\item{x}{funts object}

\item{...}{Unused additional parameters}
}
\value{
Numerics indicating the dimension of the funts object
}
\description{
Dimension of funts
}
\examples{
dim(funts(electricity))
}
