% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbow_method.R
\name{elbow_method}
\alias{elbow_method}
\title{Elbow Method}
\usage{
elbow_method(
  data,
  test_function = compute_Mn,
  trim_function = function(...) {
     0
 },
  errorType = "L2",
  ...
)
}
\arguments{
\item{data}{Numeric data.frame with rows for evaluated values and columns
indicating FD}

\item{test_function}{Function with the first argument being data. Additional
arguments as passed in as ... . This function should compute a statistic for
each candidate change point (i.e. each curve or column of data), returned as
a named value: allValues.}

\item{trim_function}{XXXXXXX}

\item{errorType}{String of 'L2' or 'Tr' indicating the error function to use}

\item{...}{Additional parameters to pass into the respective functions}
}
\value{
list with three elements.
\itemize{
\item \strong{CPInfo}: Data.frame of candidate changes, ordered by impact. The
columns are: CP, Var, and Percent. CP indicates change location, Var is
total variance, Percent is the percentage of variance removed compared
to the previous step. The first row has CP=NA, indicating the no change
scenario.
\item \strong{VarPlot}: A ggplot object showing the variance for increasing
number of changes.
\item \strong{PerPlot}: A ggplot object showing the percent of variance removed
compared to the previous step.
}
}
\description{
Method to determine the number of change points using the elbow method. Note,
cascading change points are not considered to allow for every possible
number of change points to be selectable.
}
\examples{
results <- elbow_method(
  data = electricity[, 1:50],
  trim_function = function(...) {
    10
  }
)
}
