% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fChange-package.R
\docType{package}
\name{fChange-package}
\alias{fChange}
\alias{fChange-package}
\title{General Information for fChange}
\description{
Change point detection for discretely observed, yet innately functional data. Includes methods to process and investigate such data, including functional ACF, smoothing, change point detection, plotting, and so forth.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jrvanderdoes.github.io/fChange/}
}

}
\author{
\strong{Maintainer}: Jeremy VanderDoes \email{jeremy.vanderdoes@gmail.com} (\href{https://orcid.org/0009-0001-9885-3073}{ORCID})

Authors:
\itemize{
  \item Gregory Rice
  \item Martin Wendler
}

}
\keyword{internal}
