% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fChange-package.R
\docType{package}
\name{fChange-package}
\alias{fChange}
\alias{fChange-package}
\title{General Information for fChange}
\description{
Analyze functional data and its change points. Includes functionality to store and process data, summarize and validate assumptions, characterize and perform inference of change points, and provide visualizations. Data is stored as discretely collected observations without requiring the selection of basis functions. For more details see chapter 8 of Horvath and Rice (2024) \doi{10.1007/978-3-031-51609-2}. Additional papers are forthcoming. Focused works are also included in the documentation of corresponding functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jrvanderdoes.github.io/fChange/}
}

}
\author{
\strong{Maintainer}: Jeremy VanderDoes \email{jeremy.vanderdoes@gmail.com} (\href{https://orcid.org/0009-0001-9885-3073}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Gregory Rice
  \item Martin Wendler
}

}
\keyword{internal}
