% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funts.R
\name{funts}
\alias{funts}
\title{Define funts Object}
\usage{
funts(
  X,
  labels = colnames(as.data.frame(X)),
  intraobs = seq(0, 1, length.out = nrow(X))
)
}
\arguments{
\item{labels}{Labels for the observations. Defaults to the column names of
\code{x}.}

\item{intraobs}{Vector of numerics indicating the points for each
observation. Defaults to even spacing on \link{0,1}, but may be uneven or
for any real numbers.}

\item{x}{Data.frame to convert into data}
}
\value{
funts object
}
\description{
Define funts Object
}
\examples{
# funts(electricity)
# funts(generate_brownian_motion(100, c(0,0.1,0.25,0.5,1)))
}
