% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_far1.R
\name{generate_far1}
\alias{generate_far1}
\title{Generate FAR(1) Data}
\usage{
generate_far1(N, resolution, sd = 1, dependence = 1/2, drop_first = FALSE)
}
\arguments{
\item{N}{Numeric for the number of observations.}

\item{resolution}{Numeric for resolution of data or a vector specifying the
observation points.}

\item{sd}{Numeric for standard deviation with Brownian motion.}

\item{dependence}{Numeric which indicates the dependence on the previous
curve.}

\item{drop_first}{Booolean if first values should be dropped so the data varies
at the first rather than starting at 0 (given that is the observed first point).
Note this will affect the resolution observed.}
}
\value{
dfts object of the data.
}
\description{
Function to generate data according to FAR(1) process.
}
\examples{
res <- generate_far1(20,10)
}
