% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_ou.R
\name{generate_ornstein_uhlenbeck}
\alias{generate_ornstein_uhlenbeck}
\title{Generate Data via Ornstein-Uhlenbeck Process}
\usage{
generate_ornstein_uhlenbeck(N, v, rho = 0)
}
\arguments{
\item{N}{Numeric for the number of observations.}

\item{v}{Numeric for resolution of data or a vector specifying the
observation points.}

\item{rho}{Numeric which indicates the dependence on the previous
curve.}
}
\value{
A dfts object for the generated data.
}
\description{
Generate autoregressive process with errors according the Ornstein-Uhlenbeck
process.
}
\examples{
generate_ornstein_uhlenbeck(N=100,v=20)
}
