% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfts.R
\name{lag.dfts}
\alias{lag.dfts}
\title{Lag dfts objects}
\usage{
\method{lag}{dfts}(x, k = 1, ...)
}
\arguments{
\item{x}{A dfts object. See \code{\link[=dfts]{dfts()}}.}

\item{k}{Integer for the number of lags (in units of observations).}

\item{...}{Unused additional parameters.}
}
\value{
A dfts object.
}
\description{
Compute a lagged version of a functional time series, shifting the time base
back by a given number of observations.
}
\examples{
result <- lag(electricity)
}
\seealso{
\code{\link[stats:lag]{stats::lag()}}, \code{\link[=diff.dfts]{diff.dfts()}}
}
