% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{linear_imputatation}
\alias{linear_imputatation}
\title{Impute data using linear function}
\usage{
linear_imputatation(data, evalPts = 1:nrow(data))
}
\arguments{
\item{data}{Numeric data.frame with rows for evaluated values and columns
indicating FD}

\item{evalPts}{Numeric vector indicating the evaluated points for each row
of data}
}
\value{
Numeric data.frame with rows for evaluated values and columns
indicating FD and no missing values
}
\description{
Impute data using linear function
}
\examples{
data_missing <- data.frame('FD1'=c(1:8,NA,10)+rnorm(10),
                           'FD2'=21:30+rnorm(10),
                           'FD3'=c(1:3,NA,rep(6,3),8,rep(NA,2))+rnorm(10))
evalPts <- c(1:10)
linear_imputatation(data_missing, evalPts)
}
