% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funts.R
\name{minmax}
\alias{minmax}
\alias{max.funts}
\alias{min.funts}
\title{Max / Min for funts Objects}
\usage{
\method{max}{funts}(x, type = c("Obs", "Intraobs"), na.rm = TRUE, ...)

\method{min}{funts}(x, type = c("Obs", "Intraobs"), na.rm = TRUE, ...)
}
\arguments{
\item{x}{funts object}

\item{type}{String indicating if finding for observation ('Obs', default),
or for pointwise values ('Intraobs').}

\item{na.rm}{Boolean if NA values should be removed. Defaults to TRUE.}

\item{...}{Additional parameters to stats function}
}
\value{
A funts object
}
\description{
Get the observation(s) with the min / max values. Selected as the observation
with the largest / smallest mean or the pointwise values.
}
\examples{
results <- max(funts(electricity))
results <- min(funts(electricity))
}
