% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{pca}
\alias{pca}
\alias{pca.default}
\title{Generic pca of Data}
\usage{
pca(x, ...)

\method{pca}{default}(x, ...)
}
\arguments{
\item{...}{Additional parameters to function based on data}

\item{object}{Object for computation of principle components analysis}
}
\value{
PCA data
}
\description{
Build on \code{prcomp}.
}
\examples{
pca(1:10)
}
