% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_examination.R
\name{pcaExploration}
\alias{pcaExploration}
\title{Principal Component Exploration}
\usage{
pcaExploration(X, order = 3, ...)
}
\arguments{
\item{X}{funts object}

\item{order}{Number of components to investigate}

\item{...}{Unused additional inputs}
}
\value{
List with the following elements:
\itemize{
\item Figures: List with figures on all components, summaries, reconstructed
values, and residuals
\item reconstruction: Reconstructed value from pcs
\item residuals: Difference between true and reconstruction
}
}
\description{
Principal Component Exploration
}
\examples{
results <- pcaExploration(funts(electricity))
}
\references{
Hyndman, R. J., & Shahid Ullah, M. (2007). Robust forecasting
of mortality and fertility rates: A functional data approach. Computational
Statistics & Data Analysis, 51(10), 4942–4956.
https://doi.org/10.1016/j.csda.2006.07.028
}
