% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funts.R
\name{plot.funts}
\alias{plot.funts}
\title{Plot funts}
\usage{
\method{plot}{funts}(x, ...)
}
\arguments{
\item{x}{A funts object}

\item{...}{Additional parameters to pass into plotting}
}
\value{
Plot
}
\description{
Plot funts
}
\examples{
# plot(funts(electricity))
}
