% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelation.R
\name{plot_autocovariance}
\alias{plot_autocovariance}
\title{Generate a 3D plot of the autocovariance surface of a given FTS}
\usage{
plot_autocovariance(fun.autocovariance, lag = 0, ...)
}
\arguments{
\item{fun.autocovariance}{A list obtained by calling the function
\code{obtain_autocovariance}.}

\item{lag}{An integer between 0 and \code{nlags}, indicating the lagged
autocovariance function to be plotted. By default 0.}

\item{...}{Further arguments passed to the  \code{persp} function.}
}
\value{
Graphic from \code{persp}
}
\description{
Obtain a 3D plot of the autocovariance surfaces of a given functional
time series. This visualization is useful to detect any kind of dependency
between the discretization points of the series.
}
\examples{
N <- 100
v <- seq(from = 0, to = 1, length.out = 10)
bbridge <- generate_brownian_bridge(N, v)
nlags <- 1
lagged_autocov <- .compute_autocovariance(X = bbridge,lag.max = nlags)
plot_autocovariance(lagged_autocov,1)
}
