% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fd.R
\name{plot_fd}
\alias{plot_fd}
\title{Plot functional data}
\usage{
plot_fd(
  data,
  CPs = NULL,
  eval_points = 1:nrow(data),
  plot_title = NULL,
  val_axis_title = "Value",
  eval_axis_title = "EvalRange",
  FD_axis_title = "FD Sims",
  FDReps = 1:ncol(data),
  eye = list(x = -1.5, y = -1.5, z = 1.5),
  aspectratio = list(x = 1, y = 1, z = 1),
  showticklabels = TRUE
)
}
\arguments{
\item{data}{A data.frame of evaluated fd objects (the columns being lines and
the rows the evaluated points)}

\item{CPs}{(Optional) Vectors of numeric values indicating the location of
change points. This will color each section differently. Default vallue
is NULL.}

\item{eval_points}{(Optional) An vector containing the points at which the
data was evaluated. Default is 1:nrow(data)}

\item{plot_title}{(Optional) String to title the plot. Default value is NULL.}

\item{val_axis_title}{(Optional) String to title the axis for values of
observation. Default value is 'Value'.}

\item{eval_axis_title}{(Optional) String to title the axis for the evaluation
range axis (observations in an FD object). Default value is is 'EvalRange'.}

\item{FD_axis_title}{(Optional) String to title the axis for FD observations.
Default value is 'FD Sims'.}

\item{FDReps}{(Optional) Vector of values for FD observation names. Default
value is 1:ncol(data).}

\item{eye}{(Optional) List with certain parameters to determine the view of
the resulting image. The default value is list(x = -1.5, y = -1.5, z = 1.5).}

\item{aspectratio}{(Optional) List with certain parameters to determine the
image size parameters. The default value is list(x=1, y=1, z=1).}

\item{showticklabels}{(Optional) Boolean to indicate if the tick labels should
be given in the image. Default value is TRUE.}
}
\value{
A plotly plot
}
\description{
\code{plot_fd} plots functional data either in an fd object or evaluated at certain points
}
\examples{
plot_fd(data=electricity[,1:100], CPs=c(50))
}
