% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfts.R
\name{print.dfts}
\alias{print.dfts}
\title{Print dfts objects}
\usage{
\method{print}{dfts}(x, ...)
}
\arguments{
\item{x}{A dfts object. See \code{\link[=dfts]{dfts()}}.}

\item{...}{unused parameter.}
}
\value{
No return value, called to format printing of dfts object to console.
}
\description{
Basic formatting to print a dfts object.
}
\examples{
electricity
}
