% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq.R
\name{qqplot.funts}
\alias{qqplot.funts}
\title{Generic Function to Compute QQ plot for funts}
\usage{
\method{qqplot}{funts}(x, TVE = 0.95, max.d = 3, qq.sep = TRUE, alpha = 0.05, ...)
}
\arguments{
\item{x}{funts object}

\item{TVE}{Numeric in \link{0,1} giving the total variance explained for selecting
the number of principal components.}

\item{max.d}{Max number of principal components}

\item{qq.sep}{Boolean for if the QQ plots should be seperate or stacked}

\item{alpha}{Significant}

\item{...}{}
}
\value{
ggplot2 for QQ plot
}
\description{
Creates normal QQ plots on the principal components of functional data
}
\examples{
result <- qqplot(funts(electricity))
}
