% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq.R
\name{qqplot}
\alias{qqplot}
\alias{qqplot.default}
\title{QQ Plot Generic Function}
\usage{
qqplot(x, ...)

\method{qqplot}{default}(x, ...)
}
\arguments{
\item{x}{Data to examine with qqplot}

\item{...}{Additional paramters based on the data}
}
\description{
QQ Plot Generic Function
}
\seealso{
\code{\link[=qqplot.dfts]{qqplot.dfts()}} \code{\link[=qqplot.default]{qqplot.default()}}
}
