% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfts.R
\name{quantile.dfts}
\alias{quantile.dfts}
\title{Quantile dfts}
\usage{
\method{quantile}{dfts}(x, probs = seq(0, 1, 0.25), ...)
}
\arguments{
\item{x}{A dfts object. See \code{\link[=dfts]{dfts()}}.}

\item{probs}{Numerics in [0,1] indicating the probabilities of interest.}

\item{...}{Additional parameters to pass into \code{\link[stats:quantile]{stats::quantile()}} function.}
}
\value{
Matrix with columns for each requested quantile computed pointwise.
}
\description{
Obtain the pointwise quantile information of functional data.
}
\examples{
result <- quantile(electricity)
result1 <- quantile(electricity,probs = 0.95)
}
\seealso{
\code{\link[stats:quantile]{stats::quantile()}}
}
