% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions_existing.R
\name{quantile.funts}
\alias{quantile.funts}
\title{Quantile funts}
\usage{
\method{quantile}{funts}(x, probs = seq(0, 1, 0.25), ...)
}
\arguments{
\item{x}{A funts object}

\item{...}{Additional parameters to pass into quantile function}
}
\value{
Matrix with columns for each requested quantile
}
\description{
Quantile funts
}
\examples{
# quantile(funts(electricity))
# quantile(funts(electricity),probs = 0.95)
}
