% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_val_function.R
\name{required_val_function}
\alias{required_val_function}
\title{Potential required value function}
\usage{
required_val_function(data)
}
\arguments{
\item{data}{Numeric data.frame with rows for evaluated values and columns
indicating FD}
}
\value{
Numeric value indicating necessary value
}
\description{
Potential required value function
}
\examples{
# Setup Data
data_KL <- generate_data_fd(ns = c(12,12,12),
    eigsList = list(c(3,2,1,0.5),
                    c(3,2,1,0.5),
                    c(3,2,1,0.5)),
    basesList = list(fda::create.bspline.basis(nbasis=4, norder=4),
                     fda::create.bspline.basis(nbasis=4, norder=4),
                     fda::create.bspline.basis(nbasis=4, norder=4)),
    meansList = c(-1,0,1),
    distsArray = c('Normal','Normal','Normal'),
    evals = seq(0,1,0.05),
    kappasArray = c(0,0,0))

required_val_function(data_KL)
}
\references{
Rice, G, Zhang, C. (2021), \emph{Consistency of binary
segmentation for multiple change-point estimation with functional data}
(https://www.sciencedirect.com/science/article/pii/S0167715221001905)
}
