% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar_detection.R
\name{scalarDetection}
\alias{scalarDetection}
\title{Scalar Characteristic Function Change Point Analysis}
\usage{
scalarDetection(Y, gam = 0.5, nSims = 200, alpha_val = 0.05)
}
\arguments{
\item{Y}{XXXXX}

\item{gam}{XXXXX}

\item{nSims}{XXXXX}

\item{alpha_val}{XXXXX}
}
\value{
XXXXX
}
\description{
Scalar Characteristic Function Change Point Analysis
}
\examples{
# XXXXX
}
