% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{sd}
\alias{sd}
\alias{sd.default}
\alias{var}
\alias{var.default}
\title{Generic sd/var of Data}
\usage{
sd(object, ...)

\method{sd}{default}(object, ...)

var(object, ...)

\method{var}{default}(object, ...)
}
\arguments{
\item{object}{Data to compute on}

\item{...}{Additional parameters based on the data}
}
\value{
Numeric(s) to explain the standard deviation / variance
}
\description{
Generic sd/var of Data

Generic Variance Standard Deviation
}
\examples{
sd(1:10)
var(1:10)
}
