% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{sd}
\alias{sd}
\alias{sd.default}
\alias{var}
\alias{var.default}
\title{Generic sd/var of Data}
\usage{
sd(x, ...)

\method{sd}{default}(x, ...)

var(x, ...)

\method{var}{default}(x, ...)
}
\arguments{
\item{...}{Additional parameters to function based on data}

\item{object}{Object for computation of standard deviation or variance of
a given data set.}
}
\value{
Numeric(s) to explain the standard deviation / variance
}
\description{
Generic sd/var of Data
}
\examples{
sd(1:10)
var(1:10)
}
