% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulate_size}
\alias{simulate_size}
\title{Title}
\usage{
simulate_size(
  data_length = 100,
  dep = 0,
  nSims = 500,
  path = "C:/Users/j53vande/Downloads/fChange/sims/"
)
}
\arguments{
\item{path}{}
}
\description{
Title
}
