% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{SPYUS500}
\alias{SPYUS500}
\title{S&P 500 Index Data}
\format{
\subsection{\code{SPY500}}{

A dfts object.
}
}
\usage{
SPYUS500
}
\description{
Intraday prices for the S&P500 index (SPY) for 2019 to 2023 with holidays and
weekends removed. Minutely resolution and daily observations.
}
\keyword{datasets}
