% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.funts}
\alias{summary.funts}
\title{Summary for funts Object}
\usage{
\method{summary}{funts}(object, CPs = NULL, max.lag = 20, max.d = 2, demean = FALSE, ...)
}
\arguments{
\item{object}{funts object}

\item{CPs}{CPs, if there are any. Default is NULL.}

\item{max.lag}{Max lags to consider for ACF. Default is 20.}

\item{rainbow}{Observations shown in rainbow plot. Defaults to subset,
which summarizes the values rather than showing all observations.}
}
\value{
Plot (ggplot) summarizing the data
}
\description{
Summary for funts Object
}
\examples{
summary(funts(electricity[,1:20]), max.lag=3)
# summary(funts(electricity[,1:50]), CPs=c(20,38), max.lag=3)
# summary(funts(electricity), CPs=c(50,200,300), max.lag=5, demean = TRUE)
}
