% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.funts}
\alias{summary.funts}
\title{Summary for funts Object}
\usage{
\method{summary}{funts}(object, CPs = NULL, lag.max = 20, d.max = 2, demean = FALSE, ...)
}
\arguments{
\item{object}{funts object}

\item{CPs}{CPs, if there are any. Default is NULL.}

\item{lag.max}{Max lags to consider for ACF. Default is 20.}

\item{d.max}{Max number of dimensions for QQ-plot.}

\item{demean}{Boolean if data should be demeaned based on changes and create
plots based on these residuals.}

\item{...}{Unused.}
}
\value{
Plot (ggplot) summarizing the data
}
\description{
Summary for funts Object
}
\examples{
res <- summary(funts(electricity[,1:20]), lag.max=2)
# res1 <- summary(funts(electricity[,1:20]), CPs=c(3), lag.max=2)
# summary(funts(electricity), CPs=c(50,200,300), lag.max=5, demean = TRUE)
}
