% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{temperature}
\alias{temperature}
\title{Australian Temperature Data}
\format{
\subsection{\code{temperature}}{

A dfts object.
}
}
\source{
<www.bom.gov.au>
}
\usage{
temperature
}
\description{
A list of daily minimum temperature data (1859 - 2012) from reporting station
in Sydney, Australia. Any missing data was previously, linearly,
interpolated and leap days were removed for consistency.
}
\keyword{datasets}
