% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistics_tnmn.R
\name{TnMn_resampling}
\alias{TnMn_resampling}
\title{Title}
\usage{
TnMn_resampling(
  X,
  blockSize = 1,
  iters = 1000,
  replace = FALSE,
  alpha = 0.05,
  silent = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
data_KL <- generate_data_fd(
ns = c(50, 50),
eigsList = list(
  c(3, 2, 1, 0.5),
  c(3, 3, 2)
),
basesList = list(
  fda::create.bspline.basis(nbasis = 4, norder = 4),
  fda::create.fourier.basis(nbasis = 2)
),
meansList = c(0, 0.5),
distsArray = c("Normal", "Binomial"),
evals = seq(0, 1, 0.05),
kappasArray = c(0, 0.5)
)
result <- TnMn_resampling(data_KL, M=1000)
}
