% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistics_tnmn.R
\name{TnMn_resampling}
\alias{TnMn_resampling}
\title{Resampling for Tn and Mn}
\usage{
TnMn_resampling(
  X,
  blockSize = 1,
  iters = 1000,
  replace = FALSE,
  alpha = 0.05,
  silent = FALSE,
  ...
)
}
\arguments{
\item{X}{data.frame of the data}

\item{blockSize}{Numeric for block length of the resampling}

\item{iters}{Numeric for the number of iterations for resampling}

\item{replace}{Boolean for replacement changing bootstrap to permutation}

\item{alpha}{Numeric between \link{0,1} for the significance level of alpha}

\item{silent}{Boolean for if anything should be output}

\item{...}{Additional parameters}
}
\value{
List with the following items:
'tn': Tn test statistic
'mn': Mn test statistic
'location': Estimate for the change location
'cutoff_tn': Cutoff for the Tn test statistic
'cutoff_mn': Cutoff for the Mn test statistic
'pval_tn': p-value for Tn
'pval_mn': p-value for Mn
'data_bs': Bootstrap samples of the data
}
\description{
Compute resampling statistics for Tn and Mn
}
\details{
TODO:: Make this part of the original function
}
\examples{
data_KL <- generate_data_fd(
ns = c(50, 50),
eigsList = list(
  c(3, 2, 1, 0.5),
  c(3, 3, 2)
),
basesList = list(
  fda::create.bspline.basis(nbasis = 4, norder = 4),
  fda::create.fourier.basis(nbasis = 2)
),
meansList = c(0, 0.5),
distsArray = c("Normal", "Binomial"),
evals = seq(0, 1, 0.05),
kappasArray = c(0, 0.5)
)
result <- TnMn_resampling(data_KL$data, M=1000)
}
