% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_function.R
\name{trim_function}
\alias{trim_function}
\title{Potential trim function}
\usage{
trim_function(data)
}
\arguments{
\item{data}{Numeric data.frame with rows for evaluated values and columns
indicating FD}
}
\value{
Integer value indicating the amount to trim
}
\description{
Potential trim function
}
\examples{
# Setup Data
data_KL <- generate_data_fd(ns = c(12,12,12),
    eigsList = list(c(3,2,1,0.5),
                    c(3,2,1,0.5),
                    c(3,2,1,0.5)),
    basesList = list(fda::create.bspline.basis(nbasis=4, norder=4),
                     fda::create.bspline.basis(nbasis=4, norder=4),
                     fda::create.bspline.basis(nbasis=4, norder=4)),
    meansList = c(-1,0,1),
    distsArray = c('Normal','Normal','Normal'),
    evals = seq(0,1,0.05),
    kappasArray = c(0,0,0))

trim_function(data_KL)
}
\references{
Rice, G, Zhang, C. (2021), \emph{Consistency of binary
segmentation for multiple change-point estimation with functional data}
(https://arxiv.org/pdf/2001.00093.pdf)
}
