% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{var.funts}
\alias{var.funts}
\title{Variance for funts}
\usage{
\method{var}{funts}(object, type = c("operator", "pointwise"), ...)
}
\arguments{
\item{object}{Data to compute on}

\item{type}{Character to specify if an operator ('op') or pointwise ('pw')
calculation is desired}

\item{...}{Unused parameters}
}
\value{
Vector for matrix for pointwise or operator variance
}
\description{
Variance for funts
}
\examples{
var(funts(electricity),type='pointwise')
var(funts(electricity),type='operator')
}
